Find output variable names for STICS

Helper function that return the name used as input for the STICS model with a partial match.

find_STICS_var(Var = NULL)

Arguments

Var
Character vector with a (partial) STICS output variable name

Details

The function understand regex as input.

See also

Examples

library(SticsRFiles)
find_STICS_var("lai")

#>          variable
#> 4       albedolai
#> 168        exolai
#> 243        innlai
#> 257        lai(n)
#> 258 lai_mx_av_cut
#> 259        laimax
#> 260     laisen(n)
#> 483         splai
#> 536       ulai(n)
#>                                                              details
#> 4                   albedo of the crop including soil and vegetation
#> 168              reduction factor on leaf growth due to water excess
#> 243 reduction factor on leaf growth due to NNI (nitrogen deficiency)
#> 257                                          leaf area index (table)
#> 258            LAI before cut (for cut crops , for others = lai(n) )
#> 259                                          maximum leaf area index
#> 260                      leaf area index of senescent leaves (table)
#> 483                source to sink ratio of assimilates in the leaves
#> 536                                relative development unit for LAI
#>            unit file type V1 V2
#> 4            SD   sc real  0  0
#> 168         0-1    p real  0  1
#> 243 innmin to 1    p real  0  2
#> 257      m2.m-2    p real  n  2
#> 258          SD    p real  0  1
#> 259      m2.m-2    p real  0  2
#> 260      m2.m-2    p real  n  2
#> 483          SD    p real  0  2
#> 536         0-3    p real  n  0