Generate Stics sta xml file(s) from a template or an input file

Generate Stics sta xml file(s) from a template or an input file

gen_sta_xml(
  param_table = NULL,
  sta_in_file = NULL,
  out_path = getwd(),
  stics_version = "last"
)

Arguments

param_table
a table (df, tibble) containing parameters to use (see details)

sta_in_file
file path to an XML file (optional, if not povided, uses a template from the package corresponding to stics_version)

out_path
path to an optional folder where to write the output file(s)

stics_version
the stics version to use (optional, default to last). Only used if sta_in_file= NULL, see details.

Value

an invisible xmlDocument object or a list of

Details

Please see get_stics_versions_compat() for the full list of stics versions that can be used for the argument stics_version. param_table is a data.frame with the following format:

Sta_name	zr	NH3ref	latitude	patm	aclim

climatex_sta.xml	2.5	0	49	1000	20

climatex2_sta.xml	2.8	0	49	1000	20

climatex3_sta.xml	2.2	0	49	1000	20

The first column gives the sta file name (to be generated), all following columns give the parameter value to put in the file, and each line denotes a separate sta file (for e.g. several USMs).

The first column name must contain the keyword sta or Sta or STA as a prefix to be detected (as shown in the table extract above).

Examples

if (FALSE) {

xl_path <- download_usm_xl(xl_name = "inputs_stics_example.xlsx")
sta_param_df <- read_params_table(file_path = xl_path, sheet_name = "Station")
gen_sta_xml(out_path = "/path/to/dest/dir", param_table = sta_param_df)

}