Generate from a template or modify a Stics sols or usms xmlDocument

Generate from a template or modify a Stics sols or usms xmlDocument

gen_usms_sols_doc(
  doc_type,
  xml_doc = NULL,
  nodes_nb = NULL,
  nodes_param = NULL,
  stics_version = "last",
  overwrite = F
)

Arguments

doc_type
Document type

xml_doc
The xml document

nodes_nb
The number of nodes

nodes_param
Node parameter

stics_version
Version of the STICS model

overwrite
Overwrite the document values ?

Value

An xmlDocument object

Examples

if (FALSE) {
# A newly created one, with one or more usms and
# fake parameters values
xml_doc <- SticsRFiles:::gen_xml_doc(doc_type = "usms")
xml_doc <- SticsRFiles:::gen_xml_doc(doc_type = "usms", node_nb = 3)

# With changing parameters values, from an existing document
# and a data.frame contaning usms parameters values
existing_doc <- SticsRFiles:::xmldocument("/path/to/usms.xml")

out_xml_doc <- SticsRFiles:::gen_xml_doc(doc_type = "usms",
xml_doc = existing_doc, nodes_param = param_data_frame)

}