Generating one or several usms directories from a javastics workspace content

The function creates sets of input files for one or multiple usms from usms data stored in a JavaStics workspace. For multiple usms, sets will be generated into individual folders named with usm names.. Observations files will be also copied if they are named [usm_name].obs For one usm, files will be generated either in the workspace directory or in a subdirectory.

gen_usms_xml2txt(
  javastics,
  workspace = NULL,
  out_dir = NULL,
  usm = c(),
  verbose = TRUE,
  dir_per_usm_flag = TRUE,
  check = TRUE,
  java_cmd = "java",
  javastics_path = lifecycle::deprecated(),
  workspace_path = lifecycle::deprecated(),
  target_path = lifecycle::deprecated(),
  usms_list = lifecycle::deprecated(),
  check_files = lifecycle::deprecated()
)

Arguments

javastics

Path of JavaStics.

workspace

Path of a JavaStics workspace (i.e. containing the Stics XML input files). Optional, if not provided the current workspace stored in JavaStics preferences will be used.

out_dir

The path of the directory where to create usms directories (Optional), if not provided the JavaStics workspace will be used as root

usm

List of usms to generate (Optional). If not provided, all usms contained in workspace_path/usms.xml file will be generated.

verbose

Logical value for displaying information while running

dir_per_usm_flag

logical, TRUE if one want to create one directory per USM, FALSE if USM files are generated in the target_path (only useful for usms_list of size one)

check

Logical, TRUE to check if usms files exist, FALSE otherwise

java_cmd

The java virtual machine command name or executable path

javastics_path

javastics_path is no longer supported, use javastics instead.

workspace_path

workspace_path is no longer supported, use workspace instead.

target_path

target_path is no longer supported, use out_dir instead.

usms_list

usms_list is no longer supported, use usm instead.

check_files

check_files is no longer supported, use check instead.

Value

A list with named elements: usms_path : created directories paths (for storing Stics input files), files : generated files list (in JavaStics workspace origin), copy_status : logical value vector, indicating if all files have been generated for each usm obs_copy_status : logical value vector, indicating if observaion files have been successfully copied in usms directories

Examples

if (FALSE) {
javastics <- "/path/to/JavaSTICS/folder"
workspace <- "/path/to/workspace"

# For all usms
gen_usms_xml2txt(javastics, workspace)

# For an usms list
usm <- c("usm1", "usm2")
gen_usms_xml2txt(javastics, workspace, usm)

# For one usm
gen_usms_xml2txt(javastics, workspace, dir_per_usm_flag = FALSE, "usm1")
}