Read STICS observation or simulation files (.obs or mod_s)

Read STICS observation or simulation files from a JavaStics workspace and store data into a list per usm. Used by get_obs() and get_daily_results(). Operate first computation and then call get_file_().

get_file(
  workspace,
  usm_name = NULL,
  var_list = NULL,
  doy_list = NULL,
  dates_list = NULL,
  usms_filename = NULL,
  javastics_path = NULL,
  verbose = TRUE,
  type = c("sim", "obs")
)

Arguments

workspace
Path of a JavaStics workspace, or a vector of (recursive call).

usm_name
Vector of usms to read (optional, used to filter usms)

var_list
vector of output variables names to filter (optional, see get_var_info() to get the names of the variables)

doy_list
vector of cumulative DOYs to filter (optional)

dates_list
list of dates to filter (optional, should be a POSIX date)

usms_filename
Name of the usm file (optional)

javastics_path
JavaStics installation path (optional, needed if the plant files are not in the workspace but rather in the JavaStics default workspace). Only used to get the plants names.

verbose
Logical value (optional), TRUE to display infos on error, FALSE otherwise (default)

type
The type of file to read, either "obs" or "sim".

file_name
The names of the files to read (optional). read from usms.xml file if provided, or using file names in the workspace directly

plant_names
The plants names associated to the files (optional). If not provided, read usms.xml, or use default values if not available.

Value

A named list of data.frames with observations or simulation data. The list elements are named after the usms names.

Details

The .obs files names should match USMs mames, e.g. for a usm called "banana", the .obs file should be named banana.obs. For intercrops, the name should be suffixed by "p" for the principal and "a" for the associated plant.