Getting parameters information using partial search words

Getting parameters information using partial search words

get_param_data_df(
  parameter = NULL,
  file_path = NULL,
  version = "last",
  kind = "all",
  exact = FALSE
)

Arguments

parameter
Optional name or partial name or a vector of

file_path
Optional, xml file path or a vector of

version
Optional, Stics version (default value, "last") Only the 2 last are referenced: V9.0, V9.1

kind
Kind of information to be retrieved for parameters among "parameter", "formalism" or "all" for both of them

exact
Logical, if TRUE, the exact name is searched

Value

A data.frame containing parameters names and their origin (file name), and their bounds or the formalism they belong to, or both of them.

Details

If not any name vector of names, information are extracted for all the existing parameters

Examples