Finding parameters information using partial search words

Helper function that returns names and descriptions of STICS input parameters from a partial name and/or descriptive keywords.

get_param_info(param = NULL, keyword = NULL, stics_version = "latest")

Arguments

param

Vector of parameter names (or partial names). Optional, if not provided, the function returns information for all parameters

keyword

Optional, strings or a vector of to be used for searching in parameters names and definition

stics_version

Name of the STICS version. Optional, can be used to search parameters information relative to a specific STICS version. By default the latest version returned by get_stics_versions_compat() is used.

Value

A data.frame with information about parameter(s) with columns name,file,min,max, definition

Details

The function understand regex as input.

Examples

# Find by parameter name (fuzzy search):
SticsRFiles::get_param_info("alb")
#>                      name    file  min max                  definition
#> 55                 albedo  PARSOL 0.05 0.6 albedo of the bare dry soil
#> 56   albedomulchplastique  PARTEC 0.05 0.8     albedo of plastic cover
#> 57     albedomulchresidus   PARAM 0.05 0.8       albedo of plant mulch
#> 58  albedomulchresidus(1)   PARAM 0.05 0.8       albedo of plant mulch
#> 59 albedomulchresidus(10)   PARAM 0.05 0.8       albedo of plant mulch
#> 60 albedomulchresidus(11)   PARAM 0.05 0.8       albedo of plant mulch
#> 61 albedomulchresidus(12)   PARAM 0.05 0.8       albedo of plant mulch
#> 62 albedomulchresidus(13)   PARAM 0.05 0.8       albedo of plant mulch
#> 63 albedomulchresidus(14)   PARAM 0.05 0.8       albedo of plant mulch
#> 64 albedomulchresidus(15)   PARAM 0.05 0.8       albedo of plant mulch
#> 65 albedomulchresidus(16)   PARAM 0.05 0.8       albedo of plant mulch
#> 66 albedomulchresidus(17)   PARAM 0.05 0.8       albedo of plant mulch
#> 67 albedomulchresidus(18)   PARAM 0.05 0.8       albedo of plant mulch
#> 68 albedomulchresidus(19)   PARAM 0.05 0.8       albedo of plant mulch
#> 69  albedomulchresidus(2)   PARAM 0.05 0.8       albedo of plant mulch
#> 70 albedomulchresidus(20)   PARAM 0.05 0.8       albedo of plant mulch
#> 71 albedomulchresidus(21)   PARAM 0.05 0.8       albedo of plant mulch
#> 72  albedomulchresidus(3)   PARAM 0.05 0.8       albedo of plant mulch
#> 73  albedomulchresidus(4)   PARAM 0.05 0.8       albedo of plant mulch
#> 74  albedomulchresidus(5)   PARAM 0.05 0.8       albedo of plant mulch
#> 75  albedomulchresidus(6)   PARAM 0.05 0.8       albedo of plant mulch
#> 76  albedomulchresidus(7)   PARAM 0.05 0.8       albedo of plant mulch
#> 77  albedomulchresidus(8)   PARAM 0.05 0.8       albedo of plant mulch
#> 78  albedomulchresidus(9)   PARAM 0.05 0.8       albedo of plant mulch
#> 79                 albveg STATION 0.05 0.3    albedo of the vegetation
SticsRFiles::get_param_info("alb[e]?")
#>                      name    file  min max                  definition
#> 55                 albedo  PARSOL 0.05 0.6 albedo of the bare dry soil
#> 56   albedomulchplastique  PARTEC 0.05 0.8     albedo of plastic cover
#> 57     albedomulchresidus   PARAM 0.05 0.8       albedo of plant mulch
#> 58  albedomulchresidus(1)   PARAM 0.05 0.8       albedo of plant mulch
#> 59 albedomulchresidus(10)   PARAM 0.05 0.8       albedo of plant mulch
#> 60 albedomulchresidus(11)   PARAM 0.05 0.8       albedo of plant mulch
#> 61 albedomulchresidus(12)   PARAM 0.05 0.8       albedo of plant mulch
#> 62 albedomulchresidus(13)   PARAM 0.05 0.8       albedo of plant mulch
#> 63 albedomulchresidus(14)   PARAM 0.05 0.8       albedo of plant mulch
#> 64 albedomulchresidus(15)   PARAM 0.05 0.8       albedo of plant mulch
#> 65 albedomulchresidus(16)   PARAM 0.05 0.8       albedo of plant mulch
#> 66 albedomulchresidus(17)   PARAM 0.05 0.8       albedo of plant mulch
#> 67 albedomulchresidus(18)   PARAM 0.05 0.8       albedo of plant mulch
#> 68 albedomulchresidus(19)   PARAM 0.05 0.8       albedo of plant mulch
#> 69  albedomulchresidus(2)   PARAM 0.05 0.8       albedo of plant mulch
#> 70 albedomulchresidus(20)   PARAM 0.05 0.8       albedo of plant mulch
#> 71 albedomulchresidus(21)   PARAM 0.05 0.8       albedo of plant mulch
#> 72  albedomulchresidus(3)   PARAM 0.05 0.8       albedo of plant mulch
#> 73  albedomulchresidus(4)   PARAM 0.05 0.8       albedo of plant mulch
#> 74  albedomulchresidus(5)   PARAM 0.05 0.8       albedo of plant mulch
#> 75  albedomulchresidus(6)   PARAM 0.05 0.8       albedo of plant mulch
#> 76  albedomulchresidus(7)   PARAM 0.05 0.8       albedo of plant mulch
#> 77  albedomulchresidus(8)   PARAM 0.05 0.8       albedo of plant mulch
#> 78  albedomulchresidus(9)   PARAM 0.05 0.8       albedo of plant mulch
#> 79                 albveg STATION 0.05 0.3    albedo of the vegetation

# Find by keyword (fuzzy search in parameter name and description):
SticsRFiles::get_param_info(keyword = "bdil")
#>        name   file  min max
#> 6      adil PARPLT    1   7
#> 7   adilmax PARPLT    3  10
#> 136    bdil PARPLT 0.01 0.8
#> 137 bdilmax PARPLT 0.01 0.8
#>                                                                  definition
#> 6        parameter of the critical dilution curve [Nplante]=adil MS^(-bdil)
#> 7   parameter of the maximum dilution curve [Nplante]=adilmax MS^(-bdilmax)
#> 136      parameter of the critical dilution curve [Nplante]=adil MS^(-bdil)
#> 137 parameter of the maximum dilution curve [Nplante]=adilmax MS^(-bdilmax)

# Find for a particular version:
SticsRFiles::get_param_info("alb", stics_version = "V9.0")
#>                    name    file  min max                  definition
#> 13               albedo  PARSOL 0.05 0.6 albedo of the bare dry soil
#> 14 albedomulchplastique  PARTEC 0.05 0.8     albedo of plastic cover
#> 15   albedomulchresidus   PARAM 0.05 0.8        albedo of crop mulch
#> 16               albveg STATION 0.05 0.3    albedo of the vegetation