Read STICS input parameters from text files

Read STICS model input parameters from a usm in text format (STICS input) Generally used after calling building a usm with JavaStics.

Read a specific STICS model input parameter file. Users would generally use the wrapper get_param_txt() instead.

get_param_txt(dirpath = getwd(), param = NULL, variety = NULL, ...)

get_ini_txt(filepath = "ficini.txt")

get_general_txt(filepath = "tempopar.sti")

get_tmp_txt(filepath = "tempoparv6.sti")

get_plant_txt(filepath = "ficplt1.txt", variety = NULL)

get_tec_txt(
  filepath = "fictec1.txt",
  several_fert,
  several_thin,
  is_pasture,
  ...
)

get_soil_txt(filepath = "param.sol")

get_station_txt(filepath = "station.txt")

get_usm_txt(filepath = "new_travail.usm")

Arguments

dirpath
USM directory path

param
Parameter name. Optional, if not provided, the function return an object with all parameters

variety
Integer. The plant variety to get the parameter from.

...
Further arguments to pass (for future-proofing only)

filepath
File path

several_fert
Is there several fertilization in the USM ? See details.

several_thin
Is there several thinning in the USM ? See details.

is_pasture
Is the plant a pasture ? See details.

Value

A parameter value, or if param= NULL a list of all parameters:

ini

Initialization parameters

general

General parameters

tec

Technical parameters

plant

Plant parameters

soil

Soil parameters

station

Station parameters

A list of parameters, depending on the file/function:
ini

Initialization parameters

general

General parameters

tec

Technical parameters

plant

Plant parameters

soil

Soil parameters

station

Station parameters

tmp

Temporary parameters

Details

If the variety is not given and a param is asked, the function will return the values for the variety that is simulated in the USM by checking the variete parameter in the technical file. If param is not provided by the user, the values from all varieties will be returned unless the user ask for a given variety.

several_fert, several_thin and is_pasture are read from the tmp file (tempoparv6.sti). get_param_txt() does it automatically. If you absolutely need to use directly get_tec_txt, please see example.

Note

Users would generally use get_param_txt to identify parameters names and values and pass them to other functions.

The functions are compatible with intercrops. Users generally only use get_param_txt(), which is a wrapper for all these functions.

See also

Examples

# Read the interrow distance parameter:

if (FALSE) {
library(SticsRFiles)
path = get_examples_path( file_type = "txt" )
get_param_txt(path, param='interrang')

# Getting varietal values:

# Get the leaf lifespan of the variety used in the usm:
get_param_txt(dirpath = path, param = "durvieF")
# Get the leaf lifespan of another variety available in the plant file:
get_param_txt(dirpath = path, param = "durvieF", variety = "Furio")
# To get the values for several (or all) varieties, either put all varieties:
varieties= c("Pactol", "Cherif", "Furio", "Dunia", "Volga", "Cecilia")
get_param_txt(dirpath = path, param = "durvieF", variety = varieties)
# Or get it from the output of the function returning all parameters:
get_param_txt(dirpath = path)$plant$plant1$durvieF
}

if (FALSE) {
# Read the initialisation file (ficini.txt):
library(SticsRFiles)
path = file.path(get_examples_path( file_type = "txt"), "ficini.txt")
get_ini_txt(path)

# Read the tec file directly:

# First, get the parameters from the tmp file:
tmp= get_tmp_txt(filepath = file.path(get_examples_path( file_type = "txt"), "tempoparv6.sti"))
several_fert= ifelse(tmp$option_engrais_multiple==1,TRUE,FALSE)
several_thin= ifelse(tmp$option_thinning==1,TRUE,FALSE)
is_pasture= ifelse(tmp$option_pature==1,TRUE,FALSE)

# Then, get the technical parameters:
get_tec_txt(filepath= file.path(get_examples_path( file_type = "txt"), "fictec1.txt"),
several_fert = several_fert, several_thin = several_thin, is_pasture = is_pasture)
}