Getting parameter values from xml files

Extracting parameter values for a list of xml files and parameters

get_param_xml(
  file,
  param = NULL,
  select = NULL,
  select_value = NULL,
  xml_file = lifecycle::deprecated(),
  param_name = lifecycle::deprecated(),
  value = lifecycle::deprecated(),
  ...
)

Arguments

file

Vector of the xml file paths from which parameters values must be extracted

param

Vector of parameter names. Optional, if not provided, the function returns information for all parameters.

select

node name or attribute name to use for selection (optional, default to no selection)

select_value

Vector of values used for select (see examples). Optional, should be provided only if select is provided.

xml_file

xml_file is no longer supported, use file instead.

param_name

param_name is no longer supported, use param instead.

value

value is no longer supported, use select_value instead.

...

Pass further arguments to get_param_value()

Value

A list of parameter values for each xml_file (a list of list)

Examples

if (FALSE) {
# Soil file
file <- file.path(get_examples_path(file_type = "xml"), "sols.xml")

# For all soils
get_param_xml(file)
get_param_xml(file, "argi")
get_param_xml(file, c("argi", "norg"))

# For one soil selection
get_param_xml(file, "argi", select = "sol", select_value = "solcanne")

# For soils and parameters vectors
# scalar parameters per soil
get_param_xml(file, c("argi", "norg"),
  select = "sol", select_value = c("solcanne", "solbanane")
)

# vector parameters per soil (5 values, one per soil layer)
get_param_xml(file, c("epc", "HCCF"),
  select = "sol", select_value = c("solcanne", "solbanane")
)

# Crop management file
file <- file.path(get_examples_path(file_type = "xml"), "file_tec.xml")

get_param_xml(file)

# Getting parameters for irrigation (date and quantity)
get_param_xml(file, c("julapI_or_sum_upvt", "amount"))

# Getting all parameters for a given formalism: "irrigation"
get_param_xml(file, select = "formalisme", select_value = "irrigation")
}