Get a list of Stics xml parameters names and values from a table (data.frame, tibble)

Get a list of Stics xml parameters names and values from a table (data.frame, tibble)

get_params_from_table(
  params_table,
  param_names = NULL,
  xml_doc,
  lines_id = NULL,
  stopping = FALSE,
  dict = NULL
)

Arguments

params_table
a table (df, tibble) containing parameters to use

param_names
parameters names to select

xml_doc
an xmlDocument object (of any xml type)

lines_id
table lines identifiers to select

stopping
logical value for stopping if any unknown parameters (if TRUE)

dict
List of names correspondence between short names (tags) and real parameters names

Value

a named list (with param names as list names) of data.frame/tibble

Examples

if (FALSE) {

download_usm_xl(xl_name = "inputs_stics_example.xlsx", dest_dir = "/path/to/dest/dir")
xl_path <- file.path("/path/to/dest/dir","inputs_stics_example.xlsx")
ini_param_df <- read_excel(xl_path, sheet = "Ini")
xml_path <- "path/to/ini/xml"
ini_doc <- SticsRFiles:::xmldocument(xml_path)
get_params_from_table(ini_param_df, ini_doc)

}