Extracting data from the STICS report file

Extracting data from the STICS report file

get_report_results(
  workspace,
  file_name = "mod_rapport.sti",
  usm = NULL,
  var_list = NULL,
  usm_name = lifecycle::deprecated()
)

Arguments

workspace

Path of the directory containing the STICS report file to read.

file_name

A report file name among "mod_rapport.sti" (default), "mod_rapportA.sti", "mod_rapportP.sti"

usm

Vector of USM names. Optional, if not provided, the function returns the results for all USMs.

var_list

vector of output variables names to filter (optional, see get_var_info() to get the names of the variables)

usm_name

usm_name is no longer supported, use usm instead.

Value

A data.frame

Details

The data may be filtered using usm_name vector of usm names and and/or var_list vector of variables names. In the returned data.frame, variables names respect the same syntax as in the get_sim output.

Examples

path <- get_examples_path(file_type = "sti")
get_report_results(workspace = path)
#>                 P_usm     wlieu ansemis P_iwater ancours ifin P_ichsl group
#> 1          DurumWheat  climbldj    2000      243    2001   39       1     0
#> 2          DurumWheat  climbldj    2000      243    2001  158       1     0
#> 3   DurumWheat_4years  statio5j    1992      270    1993   44       1     0
#> 4   DurumWheat_4years  statio5j    1992      270    1993  164       1     1
#> 5   DurumWheat_4years  statio5j    1993      270    1994   57       1     0
#> 6   DurumWheat_4years  statio5j    1993      270    1994  165       1     1
#> 7   DurumWheat_4years  statio5j    1994      270    1995   38       1     0
#> 8   DurumWheat_4years  statio5j    1994      270    1995  163       1     1
#> 9   DurumWheat_4years  statio5j    1995      270    1996   40       1     0
#> 10  DurumWheat_4years  statio5j    1995      270    1996  165       1     1
#> 11    DurumWheat_snow clim_snow    2009      243    2010  128       1     0
#> 12    DurumWheat_snow clim_snow    2009      243    2010  207       1     0
#> 13          SugarCane  climcanj    1998      286    1998  312       1     0
#> 14          SugarCane  climcanj    1998      286    1999  285       1    -1
#> 15           Turmeric  flux000j    2013       50    2013  180       1     0
#> 16           Turmeric  flux000j    2013       50    2013  360       1    -1
#> 17             banana  climbanj    1996       30    1996  115       1     0
#> 18             banana  climbanj    1996       30    1996  264       1     1
#> 19             barley  climorgj    2002       88    2002  146       1     0
#> 20             barley  climorgj    2002       88    2002  196       1    -1
#> 21      cc_BristleOat Auzeville    2013      234    2013  317       1     0
#> 22      cc_BristleOat Auzeville    2013      234    2013  326       1    -1
#> 23   cc_CrimsonClover Auzeville    2013      234    2013  317       1     0
#> 24   cc_CrimsonClover Auzeville    2013      234    2014   39       1    -1
#> 25 cc_ItalianRyegrass Auzeville    2012      212    2012  278       1     0
#> 26 cc_ItalianRyegrass Auzeville    2012      212    2013   33       1    -1
#> 27         cc_mustard Auzeville    2013      234    2013  254       1     0
#> 28         cc_mustard Auzeville    2013      234    2013  299       1    -1
#> 29           cc_vetch Auzeville    2012      212    2012  246       1     0
#> 30           cc_vetch Auzeville    2012      212    2012  299       1    -1
#> 31        demo_Wheat1  statio5j    1992      275    1993   82       1     0
#> 32        demo_Wheat1  statio5j    1992      275    1993  181       1     1
#> 33        demo_maize3  statio5j    1994       60    1994  155       1     0
#> 34        demo_maize3  statio5j    1994       60    1994  234       1     2
#> 35             fescue  climfetj    2004      168    2004  168       1     0
#> 36             fescue  climfetj    2004      168    2004  210       1    -1
#> 37               flax  climlinj    2002       64    2002  135       1     0
#> 38               flax  climlinj    2002       64    2002  231       1     1
#> 39              grass  prairiej    1995        1    1995   49       1     0
#> 40              grass  prairiej    1995        1    1995  151       1     0
#> 41              grass  prairiej    1995        1    1995  152       1     0
#> 42              grass  prairiej    1995        1    1995  365       1    -1
#> 43            lettuce  climsalj    2000       12    2000   43       1     0
#> 44            lettuce  climsalj    2000       12    2000   74       1    -1
#> 45              maize  climaisj    1996      112    1996  163       1     0
#> 46              maize  climaisj    1996      112    1996  300       1    -1
#> 47       maize_4years  statio5j    1992        1    1992  165       1     0
#> 48       maize_4years  statio5j    1992        1    1992  251       1     2
#> 49       maize_4years  statio5j    1993        1    1993  154       1     0
#> 50       maize_4years  statio5j    1993        1    1993  239       1     2
#> 51       maize_4years  statio5j    1994        1    1994  157       1     0
#> 52       maize_4years  statio5j    1994        1    1994  235       1     2
#> 53       maize_4years  statio5j    1995        1    1995  165       1     0
#> 54       maize_4years  statio5j    1995        1    1995  246       1     2
#> 55       maize_4years  statio5j    1996        1    1996  167       1     0
#> 56       maize_4years  statio5j    1996        1    1996  260       1     2
#> 57                pea  climpoij    2003       70    2003  110       1     0
#> 58                pea  climpoij    2003       70    2003  180       1     0
#> 59             potato  climpdtj    1997       91    1997  136       1     0
#> 60             potato  climpdtj    1997       91    1997  250       1    -1
#> 61         proto_rice  fourquej    1988      100    1988  184       1     0
#> 62         proto_rice  fourquej    1988      100    1988  255       1     1
#> 63           rapeseed  climcolj    1994      252    1994  305       1     0
#> 64           rapeseed  climcolj    1994      252    1995  185       1     0
#> 65            sorghum  climsorj    1996      112    1996  173       1     0
#> 66            sorghum  climsorj    1996      112    1996  252       1     1
#> 67            soybean  climsojj    1990      180    1990  204       1     0
#> 68            soybean  climsojj    1990      180    1990  263       1     0
#> 69         strawberry  climfraj    2001       38    2001   65       1     0
#> 70         strawberry  climfraj    2001       38    2001  133       1     2
#> 71          sugarbeet  climbetj    1994        1    1994  169       1     0
#> 72          sugarbeet  climbetj    1994        1    1994  335       1    -1
#> 73          sunflower  climtouj    1996      106    1996  155       1     0
#> 74          sunflower  climtouj    1996      106    1996  232       1     2
#> 75            timothy  chapaisj    2001       91    2001  131       1     0
#> 76            timothy  chapaisj    2001       91    2001  165       1     0
#> 77            timothy  chapaisj    2001       91    2001  171       1     0
#> 78            timothy  chapaisj    2001       91    2001  207       1     0
#> 79            timothy  chapaisj    2001       91    2001  214       1     0
#> 80            timothy  chapaisj    2001       91    2001  264       1     0
#> 81             tomato  climtomj    2000      308    2000  341       1     0
#> 82             tomato  climtomj    2000      308    2001   51       1     0
#> 83               vine  climvigj    2003      213    2004  115       1     0
#> 84               vine  climvigj    2003      213    2004  275       1     1
#> 85              wheat  climblej    1994      290    1995   83       1     0
#> 86              wheat  climblej    1994      290    1995  191       1     0
#>    P_codeplante  stade           nomversion masec_n mafruit iflos imats irecs
#> 1           ble    amf 9.2_r2519_2020-16-04   0.737   0.000     0     0     0
#> 2           ble    mat 9.2_r2519_2020-16-04  18.166   5.228   489   524     0
#> 3           ble    amf 9.2_r2519_2020-16-04   0.749   0.000     0     0     0
#> 4           ble recphy 9.2_r2519_2020-16-04  15.977   5.060   494   530   530
#> 5           ble    amf 9.2_r2519_2020-16-04   0.870   0.000     0     0     0
#> 6           ble recphy 9.2_r2519_2020-16-04  15.997   5.194   493   530   530
#> 7           ble    amf 9.2_r2519_2020-16-04   0.763   0.000     0     0     0
#> 8           ble recphy 9.2_r2519_2020-16-04  16.245   5.555   489   528   528
#> 9           ble    amf 9.2_r2519_2020-16-04   0.594   0.000     0     0     0
#> 10          ble recphy 9.2_r2519_2020-16-04  16.643   5.216   493   530   530
#> 11          ble    amf 9.2_r2519_2020-16-04   0.698   0.000     0     0     0
#> 12          ble    mat 9.2_r2519_2020-16-04  12.228   3.672   538   572     0
#> 13          can    amf 9.2_r2519_2020-16-04   0.037   0.000     0     0     0
#> 14          can recbut 9.2_r2519_2020-16-04  57.811  29.513   394   650   650
#> 15          tur    amf 9.2_r2519_2020-16-04   0.241   0.000     0     0     0
#> 16          tur recbut 9.2_r2519_2020-16-04  10.041   8.650   193   360   360
#> 17          ban    amf 9.2_r2519_2020-16-04   0.499   0.000     0     0     0
#> 18          ban recphy 9.2_r2519_2020-16-04   7.234   3.617   192   264   264
#> 19          esc    amf 9.2_r2519_2020-16-04   1.967   0.000     0     0     0
#> 20          esc recbut 9.2_r2519_2020-16-04  10.116   0.426   192   196   196
#> 21          avr    amf 9.2_r2519_2020-16-04   1.974   0.000     0     0     0
#> 22          avr recbut 9.2_r2519_2020-16-04   2.016   0.000   326   326   326
#> 23          tri    amf 9.2_r2519_2020-16-04   2.365   0.000     0     0     0
#> 24          tri recbut 9.2_r2519_2020-16-04   3.323   0.000   404   404   404
#> 25          rgi    amf 9.2_r2519_2020-16-04   1.140   0.000     0     0     0
#> 26          rgi recbut 9.2_r2519_2020-16-04   3.307   0.000   399   399   399
#> 27          mou    amf 9.2_r2519_2020-16-04   0.183   0.000     0     0     0
#> 28          mou recbut 9.2_r2519_2020-16-04   2.069   0.000   299   299   299
#> 29          ves    amf 9.2_r2519_2020-16-04   0.132   0.000     0     0     0
#> 30          ves recbut 9.2_r2519_2020-16-04   3.382   0.000   299   299   299
#> 31          ble    amf 9.2_r2519_2020-16-04   1.048   0.000     0     0     0
#> 32          ble recphy 9.2_r2519_2020-16-04  15.752   5.562   514   547   547
#> 33          mai    amf 9.2_r2519_2020-16-04   0.194   0.000     0     0     0
#> 34          mai recphy 9.2_r2519_2020-16-04   6.878   3.091   190   234   234
#> 35          fet    amf 9.2_r2519_2020-16-04   2.035   0.000     0     0     0
#> 36          fet recbut 9.2_r2519_2020-16-04   4.949   0.000   210   210   210
#> 37          lin    amf 9.2_r2519_2020-16-04   0.812   0.000     0     0     0
#> 38          lin recphy 9.2_r2519_2020-16-04   4.983   2.350   158   231   231
#> 39          fou    amf 9.2_r2519_2020-16-04   0.501   0.361     0     0     0
#> 40          fou   date 9.2_r2519_2020-16-04   6.397   6.084     0     0     1
#> 41          fou    amf 9.2_r2519_2020-16-04   0.039   0.039     0     0     1
#> 42          fou recbut 9.2_r2519_2020-16-04   2.602   2.530   365   365   365
#> 43          sal    amf 9.2_r2519_2020-16-04   0.507   0.000     0     0     0
#> 44          sal recbut 9.2_r2519_2020-16-04   0.616   0.000    74    74    74
#> 45          mai    amf 9.2_r2519_2020-16-04   0.346   0.000     0     0     0
#> 46          mai recbut 9.2_r2519_2020-16-04  24.218  14.850   211   300   300
#> 47          mai    amf 9.2_r2519_2020-16-04   0.158   0.000     0     0     0
#> 48          mai recphy 9.2_r2519_2020-16-04  14.492   9.323   204   251   251
#> 49          mai    amf 9.2_r2519_2020-16-04   0.266   0.000     0     0     0
#> 50          mai recphy 9.2_r2519_2020-16-04  13.546   8.585   191   239   239
#> 51          mai    amf 9.2_r2519_2020-16-04   0.286   0.000     0     0     0
#> 52          mai recphy 9.2_r2519_2020-16-04  12.515   8.010   191   235   235
#> 53          mai    amf 9.2_r2519_2020-16-04   0.287   0.000     0     0     0
#> 54          mai recphy 9.2_r2519_2020-16-04  12.174   7.853   199   246   246
#> 55          mai    amf 9.2_r2519_2020-16-04   0.319   0.000     0     0     0
#> 56          mai recphy 9.2_r2519_2020-16-04  16.046   9.447   203   260   260
#> 57          poi    amf 9.2_r2519_2020-16-04   0.033   0.000     0     0     0
#> 58          poi    mat 9.2_r2519_2020-16-04   5.650   2.877   143   180     0
#> 59          pdt    amf 9.2_r2519_2020-16-04   0.015   0.000     0     0     0
#> 60          pdt recbut 9.2_r2519_2020-16-04  18.008  15.307   158   250   250
#> 61          riz    amf 9.2_r2519_2020-16-04   1.870   0.000     0     0     0
#> 62          riz recphy 9.2_r2519_2020-16-04   9.479   3.432   212   255   255
#> 63          col    amf 9.2_r2519_2020-16-04   0.153   0.000     0     0     0
#> 64          col    mat 9.2_r2519_2020-16-04  12.789   5.708   494   550     0
#> 65          sor    amf 9.2_r2519_2020-16-04   0.953   0.000     0     0     0
#> 66          sor recphy 9.2_r2519_2020-16-04   8.930   6.192   215   252   252
#> 67          soj    amf 9.2_r2519_2020-16-04   0.087   0.000     0     0     0
#> 68          soj    mat 9.2_r2519_2020-16-04   6.554   2.899   231   263     0
#> 69          fra    amf 9.2_r2519_2020-16-04   0.925   0.000    61     0     0
#> 70          fra recphy 9.2_r2519_2020-16-04   3.091   1.952    61   133   133
#> 71          bet    amf 9.2_r2519_2020-16-04   0.582   0.219     0     0     0
#> 72          bet recbut 9.2_r2519_2020-16-04  31.268  25.226   254   335   335
#> 73          tou    amf 9.2_r2519_2020-16-04   0.546   0.000     0     0     0
#> 74          tou recphy 9.2_r2519_2020-16-04  12.047   4.546   198   232   232
#> 75          fou    amf 9.2_r2519_2020-16-04   0.691   0.691     0     0     0
#> 76          fou   date 9.2_r2519_2020-16-04   4.646   4.476   153    90    90
#> 77          fou    amf 9.2_r2519_2020-16-04   0.971   0.668   153    90    90
#> 78          fou   date 9.2_r2519_2020-16-04   2.913   2.026   153    90    90
#> 79          fou    amf 9.2_r2519_2020-16-04   0.806   0.454   153    90    90
#> 80          fou    mat 9.2_r2519_2020-16-04   2.624   1.789   153   264    90
#> 81          tom    amf 9.2_r2519_2020-16-04   0.636   0.000     0     0     0
#> 82          tom    mat 9.2_r2519_2020-16-04   3.798   2.106   356   417     0
#> 83          vig    amf 9.2_r2519_2020-16-04   0.029   0.000     0     0     0
#> 84          vig    mat 9.2_r2519_2020-16-04   2.162   0.000   523   640   616
#> 85          ble    amf 9.2_r2519_2020-16-04   0.821   0.000     0     0     0
#> 86          ble    mat 9.2_r2519_2020-16-04  16.698   7.504   514   556     0
#>      laimax  cprecip      cet QNplante    Qles
#> 1     1.124  293.000   70.863   48.365  44.226
#> 2     4.109  559.200  409.053  227.889  49.562
#> 3     0.991  117.000   54.237   27.909   0.123
#> 4     3.027  411.000  366.694  184.703   0.204
#> 5     0.968  413.000   86.123   32.368   1.611
#> 6     3.147  575.500  373.552  169.336   1.611
#> 7     1.035  187.000   64.466   29.772   5.043
#> 8     3.157  418.000  376.614  173.291   5.118
#> 9     1.139  373.500   72.397   35.083   2.936
#> 10    3.118  619.000  402.180  179.992   2.986
#> 11    0.689  586.365  251.707   47.052  16.059
#> 12    2.651  749.065  581.409  175.058  16.059
#> 13    0.066   76.000   78.429    2.136   0.000
#> 14    7.453 2856.000 1235.603  675.971   0.000
#> 15    0.413  233.000  131.691   10.075 169.361
#> 16    3.180 1118.000  767.356   79.212 362.551
#> 17    0.216  243.903  237.598    1.422   0.000
#> 18    2.640 1061.702  678.398  165.985 249.523
#> 19    1.048   76.000   90.219   61.374   0.390
#> 20    2.981  214.000  256.816  108.665   0.740
#> 21 1240.745  194.485  144.910   31.630   0.000
#> 22 1240.912  259.487  149.823   33.559   0.833
#> 23    3.691  194.500  129.519   91.976   0.000
#> 24    5.174  412.500  188.726  118.066  12.982
#> 25    1.902  100.500  108.344   23.417   0.000
#> 26    1.918  411.922  194.028   90.218   1.988
#> 27    0.477   75.000   43.187    6.324   0.000
#> 28    2.319  131.736  127.128   29.473   0.000
#> 29    0.273   74.500   44.918    4.438   0.000
#> 30    0.909  128.981  147.062  106.334   0.000
#> 31    1.088  163.500  114.404   66.818   0.000
#> 32    4.870  424.500  428.653  190.862   0.000
#> 33    0.410   68.500   80.975    5.100   0.000
#> 34    2.415  462.961  424.615   62.875   0.000
#> 35    0.800    0.000    6.100   32.014   0.000
#> 36    4.824   38.400  151.601  104.012   0.000
#> 37    1.451   87.500   93.371   27.543   0.377
#> 38    5.956  343.500  370.582   77.117   0.377
#> 39    0.471  139.000   32.600   68.692   4.266
#> 40    1.990  314.093  295.151  107.884   5.217
#> 41    1.990  314.093  298.293    0.459   5.217
#> 42    1.990  735.630  708.425   53.566   5.217
#> 43    0.000   70.000   19.551    1.291  45.868
#> 44    0.000   86.500   60.659   16.100  45.868
#> 45    0.528  110.000   54.766   11.393   1.449
#> 46    5.800  668.500  627.275  321.271   2.778
#> 47    0.307  184.000   92.497    3.794   0.509
#> 48    3.982  500.500  482.760  177.436   0.571
#> 49    0.415   74.000   65.965    7.341   1.026
#> 50    3.784  459.000  473.195  162.923   1.026
#> 51    0.497   68.500   72.257    8.076   1.263
#> 52    3.640  388.000  451.393  167.745   1.263
#> 53    0.448   72.000   71.400    9.797   0.172
#> 54    3.872  338.000  449.629  159.248   0.172
#> 55    0.536  103.000   76.095    9.264   2.218
#> 56    4.067  413.000  494.412  192.872   2.218
#> 57    0.221   51.600   57.756    1.052   0.000
#> 58    2.337  217.200  254.539  134.513   0.000
#> 59    0.035  108.500   52.282    1.119   0.040
#> 60    3.636  408.100  446.499  224.601   0.152
#> 61    1.338  428.300  195.925   76.055   0.351
#> 62    4.633  783.800  546.651  122.520   0.495
#> 63    0.191  110.900   60.849    9.703  21.583
#> 64    3.533  658.596  490.567  189.939  65.143
#> 65    1.442  124.000   77.420   39.126   1.315
#> 66    6.034  206.500  290.520  161.339   1.315
#> 67    0.130   71.000   73.222    4.997   0.000
#> 68    4.523  152.069  343.174   92.942   0.000
#> 69    0.214   20.000   11.245   48.731   0.042
#> 70    0.519   82.698   47.066   63.499   0.042
#> 71    0.574   98.500  108.474   34.034   5.100
#> 72    5.367  474.149  602.842  236.601   5.100
#> 73    0.554  156.500  103.623   17.189   0.001
#> 74    3.375  370.500  435.968  156.673   0.001
#> 75    1.210   48.400   54.436   42.351   0.037
#> 76    4.560  180.025  163.064  105.433   0.178
#> 77    4.560  196.692  182.669   28.454   0.178
#> 78    4.560  352.661  289.534   92.862   0.369
#> 79    4.560  353.523  304.711   38.886   0.369
#> 80    4.560  498.481  414.622   61.890   0.369
#> 81    0.149   59.265   67.316    5.299   0.000
#> 82    0.534  393.330  394.860  120.420   0.002
#> 83    0.024  513.600  196.661   21.383  27.396
#> 84    0.963  726.000  421.707   45.191  28.211
#> 85    0.987  303.000  104.504   40.375   0.033
#> 86    4.765  619.200  396.015  212.760   0.104

get_report_results(workspace = path, usm = c("DurumWheat", "grass"))
#>        P_usm    wlieu ansemis P_iwater ancours ifin P_ichsl group P_codeplante
#> 1 DurumWheat climbldj    2000      243    2001   39       1     0          ble
#> 2 DurumWheat climbldj    2000      243    2001  158       1     0          ble
#> 3      grass prairiej    1995        1    1995   49       1     0          fou
#> 4      grass prairiej    1995        1    1995  151       1     0          fou
#> 5      grass prairiej    1995        1    1995  152       1     0          fou
#> 6      grass prairiej    1995        1    1995  365       1    -1          fou
#>    stade           nomversion masec_n mafruit iflos imats irecs laimax cprecip
#> 1    amf 9.2_r2519_2020-16-04   0.737   0.000     0     0     0  1.124 293.000
#> 2    mat 9.2_r2519_2020-16-04  18.166   5.228   489   524     0  4.109 559.200
#> 3    amf 9.2_r2519_2020-16-04   0.501   0.361     0     0     0  0.471 139.000
#> 4   date 9.2_r2519_2020-16-04   6.397   6.084     0     0     1  1.990 314.093
#> 5    amf 9.2_r2519_2020-16-04   0.039   0.039     0     0     1  1.990 314.093
#> 6 recbut 9.2_r2519_2020-16-04   2.602   2.530   365   365   365  1.990 735.630
#>       cet QNplante   Qles
#> 1  70.863   48.365 44.226
#> 2 409.053  227.889 49.562
#> 3  32.600   68.692  4.266
#> 4 295.151  107.884  5.217
#> 5 298.293    0.459  5.217
#> 6 708.425   53.566  5.217

get_report_results(workspace = path, var_list = c("masec(n)", "QNplante"))
#>                 P_usm     wlieu ansemis P_iwater ancours ifin P_ichsl group
#> 1          DurumWheat  climbldj    2000      243    2001   39       1     0
#> 2          DurumWheat  climbldj    2000      243    2001  158       1     0
#> 3   DurumWheat_4years  statio5j    1992      270    1993   44       1     0
#> 4   DurumWheat_4years  statio5j    1992      270    1993  164       1     1
#> 5   DurumWheat_4years  statio5j    1993      270    1994   57       1     0
#> 6   DurumWheat_4years  statio5j    1993      270    1994  165       1     1
#> 7   DurumWheat_4years  statio5j    1994      270    1995   38       1     0
#> 8   DurumWheat_4years  statio5j    1994      270    1995  163       1     1
#> 9   DurumWheat_4years  statio5j    1995      270    1996   40       1     0
#> 10  DurumWheat_4years  statio5j    1995      270    1996  165       1     1
#> 11    DurumWheat_snow clim_snow    2009      243    2010  128       1     0
#> 12    DurumWheat_snow clim_snow    2009      243    2010  207       1     0
#> 13          SugarCane  climcanj    1998      286    1998  312       1     0
#> 14          SugarCane  climcanj    1998      286    1999  285       1    -1
#> 15           Turmeric  flux000j    2013       50    2013  180       1     0
#> 16           Turmeric  flux000j    2013       50    2013  360       1    -1
#> 17             banana  climbanj    1996       30    1996  115       1     0
#> 18             banana  climbanj    1996       30    1996  264       1     1
#> 19             barley  climorgj    2002       88    2002  146       1     0
#> 20             barley  climorgj    2002       88    2002  196       1    -1
#> 21      cc_BristleOat Auzeville    2013      234    2013  317       1     0
#> 22      cc_BristleOat Auzeville    2013      234    2013  326       1    -1
#> 23   cc_CrimsonClover Auzeville    2013      234    2013  317       1     0
#> 24   cc_CrimsonClover Auzeville    2013      234    2014   39       1    -1
#> 25 cc_ItalianRyegrass Auzeville    2012      212    2012  278       1     0
#> 26 cc_ItalianRyegrass Auzeville    2012      212    2013   33       1    -1
#> 27         cc_mustard Auzeville    2013      234    2013  254       1     0
#> 28         cc_mustard Auzeville    2013      234    2013  299       1    -1
#> 29           cc_vetch Auzeville    2012      212    2012  246       1     0
#> 30           cc_vetch Auzeville    2012      212    2012  299       1    -1
#> 31        demo_Wheat1  statio5j    1992      275    1993   82       1     0
#> 32        demo_Wheat1  statio5j    1992      275    1993  181       1     1
#> 33        demo_maize3  statio5j    1994       60    1994  155       1     0
#> 34        demo_maize3  statio5j    1994       60    1994  234       1     2
#> 35             fescue  climfetj    2004      168    2004  168       1     0
#> 36             fescue  climfetj    2004      168    2004  210       1    -1
#> 37               flax  climlinj    2002       64    2002  135       1     0
#> 38               flax  climlinj    2002       64    2002  231       1     1
#> 39              grass  prairiej    1995        1    1995   49       1     0
#> 40              grass  prairiej    1995        1    1995  151       1     0
#> 41              grass  prairiej    1995        1    1995  152       1     0
#> 42              grass  prairiej    1995        1    1995  365       1    -1
#> 43            lettuce  climsalj    2000       12    2000   43       1     0
#> 44            lettuce  climsalj    2000       12    2000   74       1    -1
#> 45              maize  climaisj    1996      112    1996  163       1     0
#> 46              maize  climaisj    1996      112    1996  300       1    -1
#> 47       maize_4years  statio5j    1992        1    1992  165       1     0
#> 48       maize_4years  statio5j    1992        1    1992  251       1     2
#> 49       maize_4years  statio5j    1993        1    1993  154       1     0
#> 50       maize_4years  statio5j    1993        1    1993  239       1     2
#> 51       maize_4years  statio5j    1994        1    1994  157       1     0
#> 52       maize_4years  statio5j    1994        1    1994  235       1     2
#> 53       maize_4years  statio5j    1995        1    1995  165       1     0
#> 54       maize_4years  statio5j    1995        1    1995  246       1     2
#> 55       maize_4years  statio5j    1996        1    1996  167       1     0
#> 56       maize_4years  statio5j    1996        1    1996  260       1     2
#> 57                pea  climpoij    2003       70    2003  110       1     0
#> 58                pea  climpoij    2003       70    2003  180       1     0
#> 59             potato  climpdtj    1997       91    1997  136       1     0
#> 60             potato  climpdtj    1997       91    1997  250       1    -1
#> 61         proto_rice  fourquej    1988      100    1988  184       1     0
#> 62         proto_rice  fourquej    1988      100    1988  255       1     1
#> 63           rapeseed  climcolj    1994      252    1994  305       1     0
#> 64           rapeseed  climcolj    1994      252    1995  185       1     0
#> 65            sorghum  climsorj    1996      112    1996  173       1     0
#> 66            sorghum  climsorj    1996      112    1996  252       1     1
#> 67            soybean  climsojj    1990      180    1990  204       1     0
#> 68            soybean  climsojj    1990      180    1990  263       1     0
#> 69         strawberry  climfraj    2001       38    2001   65       1     0
#> 70         strawberry  climfraj    2001       38    2001  133       1     2
#> 71          sugarbeet  climbetj    1994        1    1994  169       1     0
#> 72          sugarbeet  climbetj    1994        1    1994  335       1    -1
#> 73          sunflower  climtouj    1996      106    1996  155       1     0
#> 74          sunflower  climtouj    1996      106    1996  232       1     2
#> 75            timothy  chapaisj    2001       91    2001  131       1     0
#> 76            timothy  chapaisj    2001       91    2001  165       1     0
#> 77            timothy  chapaisj    2001       91    2001  171       1     0
#> 78            timothy  chapaisj    2001       91    2001  207       1     0
#> 79            timothy  chapaisj    2001       91    2001  214       1     0
#> 80            timothy  chapaisj    2001       91    2001  264       1     0
#> 81             tomato  climtomj    2000      308    2000  341       1     0
#> 82             tomato  climtomj    2000      308    2001   51       1     0
#> 83               vine  climvigj    2003      213    2004  115       1     0
#> 84               vine  climvigj    2003      213    2004  275       1     1
#> 85              wheat  climblej    1994      290    1995   83       1     0
#> 86              wheat  climblej    1994      290    1995  191       1     0
#>    P_codeplante  stade           nomversion masec_n QNplante
#> 1           ble    amf 9.2_r2519_2020-16-04   0.737   48.365
#> 2           ble    mat 9.2_r2519_2020-16-04  18.166  227.889
#> 3           ble    amf 9.2_r2519_2020-16-04   0.749   27.909
#> 4           ble recphy 9.2_r2519_2020-16-04  15.977  184.703
#> 5           ble    amf 9.2_r2519_2020-16-04   0.870   32.368
#> 6           ble recphy 9.2_r2519_2020-16-04  15.997  169.336
#> 7           ble    amf 9.2_r2519_2020-16-04   0.763   29.772
#> 8           ble recphy 9.2_r2519_2020-16-04  16.245  173.291
#> 9           ble    amf 9.2_r2519_2020-16-04   0.594   35.083
#> 10          ble recphy 9.2_r2519_2020-16-04  16.643  179.992
#> 11          ble    amf 9.2_r2519_2020-16-04   0.698   47.052
#> 12          ble    mat 9.2_r2519_2020-16-04  12.228  175.058
#> 13          can    amf 9.2_r2519_2020-16-04   0.037    2.136
#> 14          can recbut 9.2_r2519_2020-16-04  57.811  675.971
#> 15          tur    amf 9.2_r2519_2020-16-04   0.241   10.075
#> 16          tur recbut 9.2_r2519_2020-16-04  10.041   79.212
#> 17          ban    amf 9.2_r2519_2020-16-04   0.499    1.422
#> 18          ban recphy 9.2_r2519_2020-16-04   7.234  165.985
#> 19          esc    amf 9.2_r2519_2020-16-04   1.967   61.374
#> 20          esc recbut 9.2_r2519_2020-16-04  10.116  108.665
#> 21          avr    amf 9.2_r2519_2020-16-04   1.974   31.630
#> 22          avr recbut 9.2_r2519_2020-16-04   2.016   33.559
#> 23          tri    amf 9.2_r2519_2020-16-04   2.365   91.976
#> 24          tri recbut 9.2_r2519_2020-16-04   3.323  118.066
#> 25          rgi    amf 9.2_r2519_2020-16-04   1.140   23.417
#> 26          rgi recbut 9.2_r2519_2020-16-04   3.307   90.218
#> 27          mou    amf 9.2_r2519_2020-16-04   0.183    6.324
#> 28          mou recbut 9.2_r2519_2020-16-04   2.069   29.473
#> 29          ves    amf 9.2_r2519_2020-16-04   0.132    4.438
#> 30          ves recbut 9.2_r2519_2020-16-04   3.382  106.334
#> 31          ble    amf 9.2_r2519_2020-16-04   1.048   66.818
#> 32          ble recphy 9.2_r2519_2020-16-04  15.752  190.862
#> 33          mai    amf 9.2_r2519_2020-16-04   0.194    5.100
#> 34          mai recphy 9.2_r2519_2020-16-04   6.878   62.875
#> 35          fet    amf 9.2_r2519_2020-16-04   2.035   32.014
#> 36          fet recbut 9.2_r2519_2020-16-04   4.949  104.012
#> 37          lin    amf 9.2_r2519_2020-16-04   0.812   27.543
#> 38          lin recphy 9.2_r2519_2020-16-04   4.983   77.117
#> 39          fou    amf 9.2_r2519_2020-16-04   0.501   68.692
#> 40          fou   date 9.2_r2519_2020-16-04   6.397  107.884
#> 41          fou    amf 9.2_r2519_2020-16-04   0.039    0.459
#> 42          fou recbut 9.2_r2519_2020-16-04   2.602   53.566
#> 43          sal    amf 9.2_r2519_2020-16-04   0.507    1.291
#> 44          sal recbut 9.2_r2519_2020-16-04   0.616   16.100
#> 45          mai    amf 9.2_r2519_2020-16-04   0.346   11.393
#> 46          mai recbut 9.2_r2519_2020-16-04  24.218  321.271
#> 47          mai    amf 9.2_r2519_2020-16-04   0.158    3.794
#> 48          mai recphy 9.2_r2519_2020-16-04  14.492  177.436
#> 49          mai    amf 9.2_r2519_2020-16-04   0.266    7.341
#> 50          mai recphy 9.2_r2519_2020-16-04  13.546  162.923
#> 51          mai    amf 9.2_r2519_2020-16-04   0.286    8.076
#> 52          mai recphy 9.2_r2519_2020-16-04  12.515  167.745
#> 53          mai    amf 9.2_r2519_2020-16-04   0.287    9.797
#> 54          mai recphy 9.2_r2519_2020-16-04  12.174  159.248
#> 55          mai    amf 9.2_r2519_2020-16-04   0.319    9.264
#> 56          mai recphy 9.2_r2519_2020-16-04  16.046  192.872
#> 57          poi    amf 9.2_r2519_2020-16-04   0.033    1.052
#> 58          poi    mat 9.2_r2519_2020-16-04   5.650  134.513
#> 59          pdt    amf 9.2_r2519_2020-16-04   0.015    1.119
#> 60          pdt recbut 9.2_r2519_2020-16-04  18.008  224.601
#> 61          riz    amf 9.2_r2519_2020-16-04   1.870   76.055
#> 62          riz recphy 9.2_r2519_2020-16-04   9.479  122.520
#> 63          col    amf 9.2_r2519_2020-16-04   0.153    9.703
#> 64          col    mat 9.2_r2519_2020-16-04  12.789  189.939
#> 65          sor    amf 9.2_r2519_2020-16-04   0.953   39.126
#> 66          sor recphy 9.2_r2519_2020-16-04   8.930  161.339
#> 67          soj    amf 9.2_r2519_2020-16-04   0.087    4.997
#> 68          soj    mat 9.2_r2519_2020-16-04   6.554   92.942
#> 69          fra    amf 9.2_r2519_2020-16-04   0.925   48.731
#> 70          fra recphy 9.2_r2519_2020-16-04   3.091   63.499
#> 71          bet    amf 9.2_r2519_2020-16-04   0.582   34.034
#> 72          bet recbut 9.2_r2519_2020-16-04  31.268  236.601
#> 73          tou    amf 9.2_r2519_2020-16-04   0.546   17.189
#> 74          tou recphy 9.2_r2519_2020-16-04  12.047  156.673
#> 75          fou    amf 9.2_r2519_2020-16-04   0.691   42.351
#> 76          fou   date 9.2_r2519_2020-16-04   4.646  105.433
#> 77          fou    amf 9.2_r2519_2020-16-04   0.971   28.454
#> 78          fou   date 9.2_r2519_2020-16-04   2.913   92.862
#> 79          fou    amf 9.2_r2519_2020-16-04   0.806   38.886
#> 80          fou    mat 9.2_r2519_2020-16-04   2.624   61.890
#> 81          tom    amf 9.2_r2519_2020-16-04   0.636    5.299
#> 82          tom    mat 9.2_r2519_2020-16-04   3.798  120.420
#> 83          vig    amf 9.2_r2519_2020-16-04   0.029   21.383
#> 84          vig    mat 9.2_r2519_2020-16-04   2.162   45.191
#> 85          ble    amf 9.2_r2519_2020-16-04   0.821   40.375
#> 86          ble    mat 9.2_r2519_2020-16-04  16.698  212.760

get_report_results(workspace = path, usm = c("DurumWheat", "grass"))
#>        P_usm    wlieu ansemis P_iwater ancours ifin P_ichsl group P_codeplante
#> 1 DurumWheat climbldj    2000      243    2001   39       1     0          ble
#> 2 DurumWheat climbldj    2000      243    2001  158       1     0          ble
#> 3      grass prairiej    1995        1    1995   49       1     0          fou
#> 4      grass prairiej    1995        1    1995  151       1     0          fou
#> 5      grass prairiej    1995        1    1995  152       1     0          fou
#> 6      grass prairiej    1995        1    1995  365       1    -1          fou
#>    stade           nomversion masec_n mafruit iflos imats irecs laimax cprecip
#> 1    amf 9.2_r2519_2020-16-04   0.737   0.000     0     0     0  1.124 293.000
#> 2    mat 9.2_r2519_2020-16-04  18.166   5.228   489   524     0  4.109 559.200
#> 3    amf 9.2_r2519_2020-16-04   0.501   0.361     0     0     0  0.471 139.000
#> 4   date 9.2_r2519_2020-16-04   6.397   6.084     0     0     1  1.990 314.093
#> 5    amf 9.2_r2519_2020-16-04   0.039   0.039     0     0     1  1.990 314.093
#> 6 recbut 9.2_r2519_2020-16-04   2.602   2.530   365   365   365  1.990 735.630
#>       cet QNplante   Qles
#> 1  70.863   48.365 44.226
#> 2 409.053  227.889 49.562
#> 3  32.600   68.692  4.266
#> 4 295.151  107.884  5.217
#> 5 298.293    0.459  5.217
#> 6 708.425   53.566  5.217

get_report_results(workspace = path)
#>                 P_usm     wlieu ansemis P_iwater ancours ifin P_ichsl group
#> 1          DurumWheat  climbldj    2000      243    2001   39       1     0
#> 2          DurumWheat  climbldj    2000      243    2001  158       1     0
#> 3   DurumWheat_4years  statio5j    1992      270    1993   44       1     0
#> 4   DurumWheat_4years  statio5j    1992      270    1993  164       1     1
#> 5   DurumWheat_4years  statio5j    1993      270    1994   57       1     0
#> 6   DurumWheat_4years  statio5j    1993      270    1994  165       1     1
#> 7   DurumWheat_4years  statio5j    1994      270    1995   38       1     0
#> 8   DurumWheat_4years  statio5j    1994      270    1995  163       1     1
#> 9   DurumWheat_4years  statio5j    1995      270    1996   40       1     0
#> 10  DurumWheat_4years  statio5j    1995      270    1996  165       1     1
#> 11    DurumWheat_snow clim_snow    2009      243    2010  128       1     0
#> 12    DurumWheat_snow clim_snow    2009      243    2010  207       1     0
#> 13          SugarCane  climcanj    1998      286    1998  312       1     0
#> 14          SugarCane  climcanj    1998      286    1999  285       1    -1
#> 15           Turmeric  flux000j    2013       50    2013  180       1     0
#> 16           Turmeric  flux000j    2013       50    2013  360       1    -1
#> 17             banana  climbanj    1996       30    1996  115       1     0
#> 18             banana  climbanj    1996       30    1996  264       1     1
#> 19             barley  climorgj    2002       88    2002  146       1     0
#> 20             barley  climorgj    2002       88    2002  196       1    -1
#> 21      cc_BristleOat Auzeville    2013      234    2013  317       1     0
#> 22      cc_BristleOat Auzeville    2013      234    2013  326       1    -1
#> 23   cc_CrimsonClover Auzeville    2013      234    2013  317       1     0
#> 24   cc_CrimsonClover Auzeville    2013      234    2014   39       1    -1
#> 25 cc_ItalianRyegrass Auzeville    2012      212    2012  278       1     0
#> 26 cc_ItalianRyegrass Auzeville    2012      212    2013   33       1    -1
#> 27         cc_mustard Auzeville    2013      234    2013  254       1     0
#> 28         cc_mustard Auzeville    2013      234    2013  299       1    -1
#> 29           cc_vetch Auzeville    2012      212    2012  246       1     0
#> 30           cc_vetch Auzeville    2012      212    2012  299       1    -1
#> 31        demo_Wheat1  statio5j    1992      275    1993   82       1     0
#> 32        demo_Wheat1  statio5j    1992      275    1993  181       1     1
#> 33        demo_maize3  statio5j    1994       60    1994  155       1     0
#> 34        demo_maize3  statio5j    1994       60    1994  234       1     2
#> 35             fescue  climfetj    2004      168    2004  168       1     0
#> 36             fescue  climfetj    2004      168    2004  210       1    -1
#> 37               flax  climlinj    2002       64    2002  135       1     0
#> 38               flax  climlinj    2002       64    2002  231       1     1
#> 39              grass  prairiej    1995        1    1995   49       1     0
#> 40              grass  prairiej    1995        1    1995  151       1     0
#> 41              grass  prairiej    1995        1    1995  152       1     0
#> 42              grass  prairiej    1995        1    1995  365       1    -1
#> 43            lettuce  climsalj    2000       12    2000   43       1     0
#> 44            lettuce  climsalj    2000       12    2000   74       1    -1
#> 45              maize  climaisj    1996      112    1996  163       1     0
#> 46              maize  climaisj    1996      112    1996  300       1    -1
#> 47       maize_4years  statio5j    1992        1    1992  165       1     0
#> 48       maize_4years  statio5j    1992        1    1992  251       1     2
#> 49       maize_4years  statio5j    1993        1    1993  154       1     0
#> 50       maize_4years  statio5j    1993        1    1993  239       1     2
#> 51       maize_4years  statio5j    1994        1    1994  157       1     0
#> 52       maize_4years  statio5j    1994        1    1994  235       1     2
#> 53       maize_4years  statio5j    1995        1    1995  165       1     0
#> 54       maize_4years  statio5j    1995        1    1995  246       1     2
#> 55       maize_4years  statio5j    1996        1    1996  167       1     0
#> 56       maize_4years  statio5j    1996        1    1996  260       1     2
#> 57                pea  climpoij    2003       70    2003  110       1     0
#> 58                pea  climpoij    2003       70    2003  180       1     0
#> 59             potato  climpdtj    1997       91    1997  136       1     0
#> 60             potato  climpdtj    1997       91    1997  250       1    -1
#> 61         proto_rice  fourquej    1988      100    1988  184       1     0
#> 62         proto_rice  fourquej    1988      100    1988  255       1     1
#> 63           rapeseed  climcolj    1994      252    1994  305       1     0
#> 64           rapeseed  climcolj    1994      252    1995  185       1     0
#> 65            sorghum  climsorj    1996      112    1996  173       1     0
#> 66            sorghum  climsorj    1996      112    1996  252       1     1
#> 67            soybean  climsojj    1990      180    1990  204       1     0
#> 68            soybean  climsojj    1990      180    1990  263       1     0
#> 69         strawberry  climfraj    2001       38    2001   65       1     0
#> 70         strawberry  climfraj    2001       38    2001  133       1     2
#> 71          sugarbeet  climbetj    1994        1    1994  169       1     0
#> 72          sugarbeet  climbetj    1994        1    1994  335       1    -1
#> 73          sunflower  climtouj    1996      106    1996  155       1     0
#> 74          sunflower  climtouj    1996      106    1996  232       1     2
#> 75            timothy  chapaisj    2001       91    2001  131       1     0
#> 76            timothy  chapaisj    2001       91    2001  165       1     0
#> 77            timothy  chapaisj    2001       91    2001  171       1     0
#> 78            timothy  chapaisj    2001       91    2001  207       1     0
#> 79            timothy  chapaisj    2001       91    2001  214       1     0
#> 80            timothy  chapaisj    2001       91    2001  264       1     0
#> 81             tomato  climtomj    2000      308    2000  341       1     0
#> 82             tomato  climtomj    2000      308    2001   51       1     0
#> 83               vine  climvigj    2003      213    2004  115       1     0
#> 84               vine  climvigj    2003      213    2004  275       1     1
#> 85              wheat  climblej    1994      290    1995   83       1     0
#> 86              wheat  climblej    1994      290    1995  191       1     0
#>    P_codeplante  stade           nomversion masec_n mafruit iflos imats irecs
#> 1           ble    amf 9.2_r2519_2020-16-04   0.737   0.000     0     0     0
#> 2           ble    mat 9.2_r2519_2020-16-04  18.166   5.228   489   524     0
#> 3           ble    amf 9.2_r2519_2020-16-04   0.749   0.000     0     0     0
#> 4           ble recphy 9.2_r2519_2020-16-04  15.977   5.060   494   530   530
#> 5           ble    amf 9.2_r2519_2020-16-04   0.870   0.000     0     0     0
#> 6           ble recphy 9.2_r2519_2020-16-04  15.997   5.194   493   530   530
#> 7           ble    amf 9.2_r2519_2020-16-04   0.763   0.000     0     0     0
#> 8           ble recphy 9.2_r2519_2020-16-04  16.245   5.555   489   528   528
#> 9           ble    amf 9.2_r2519_2020-16-04   0.594   0.000     0     0     0
#> 10          ble recphy 9.2_r2519_2020-16-04  16.643   5.216   493   530   530
#> 11          ble    amf 9.2_r2519_2020-16-04   0.698   0.000     0     0     0
#> 12          ble    mat 9.2_r2519_2020-16-04  12.228   3.672   538   572     0
#> 13          can    amf 9.2_r2519_2020-16-04   0.037   0.000     0     0     0
#> 14          can recbut 9.2_r2519_2020-16-04  57.811  29.513   394   650   650
#> 15          tur    amf 9.2_r2519_2020-16-04   0.241   0.000     0     0     0
#> 16          tur recbut 9.2_r2519_2020-16-04  10.041   8.650   193   360   360
#> 17          ban    amf 9.2_r2519_2020-16-04   0.499   0.000     0     0     0
#> 18          ban recphy 9.2_r2519_2020-16-04   7.234   3.617   192   264   264
#> 19          esc    amf 9.2_r2519_2020-16-04   1.967   0.000     0     0     0
#> 20          esc recbut 9.2_r2519_2020-16-04  10.116   0.426   192   196   196
#> 21          avr    amf 9.2_r2519_2020-16-04   1.974   0.000     0     0     0
#> 22          avr recbut 9.2_r2519_2020-16-04   2.016   0.000   326   326   326
#> 23          tri    amf 9.2_r2519_2020-16-04   2.365   0.000     0     0     0
#> 24          tri recbut 9.2_r2519_2020-16-04   3.323   0.000   404   404   404
#> 25          rgi    amf 9.2_r2519_2020-16-04   1.140   0.000     0     0     0
#> 26          rgi recbut 9.2_r2519_2020-16-04   3.307   0.000   399   399   399
#> 27          mou    amf 9.2_r2519_2020-16-04   0.183   0.000     0     0     0
#> 28          mou recbut 9.2_r2519_2020-16-04   2.069   0.000   299   299   299
#> 29          ves    amf 9.2_r2519_2020-16-04   0.132   0.000     0     0     0
#> 30          ves recbut 9.2_r2519_2020-16-04   3.382   0.000   299   299   299
#> 31          ble    amf 9.2_r2519_2020-16-04   1.048   0.000     0     0     0
#> 32          ble recphy 9.2_r2519_2020-16-04  15.752   5.562   514   547   547
#> 33          mai    amf 9.2_r2519_2020-16-04   0.194   0.000     0     0     0
#> 34          mai recphy 9.2_r2519_2020-16-04   6.878   3.091   190   234   234
#> 35          fet    amf 9.2_r2519_2020-16-04   2.035   0.000     0     0     0
#> 36          fet recbut 9.2_r2519_2020-16-04   4.949   0.000   210   210   210
#> 37          lin    amf 9.2_r2519_2020-16-04   0.812   0.000     0     0     0
#> 38          lin recphy 9.2_r2519_2020-16-04   4.983   2.350   158   231   231
#> 39          fou    amf 9.2_r2519_2020-16-04   0.501   0.361     0     0     0
#> 40          fou   date 9.2_r2519_2020-16-04   6.397   6.084     0     0     1
#> 41          fou    amf 9.2_r2519_2020-16-04   0.039   0.039     0     0     1
#> 42          fou recbut 9.2_r2519_2020-16-04   2.602   2.530   365   365   365
#> 43          sal    amf 9.2_r2519_2020-16-04   0.507   0.000     0     0     0
#> 44          sal recbut 9.2_r2519_2020-16-04   0.616   0.000    74    74    74
#> 45          mai    amf 9.2_r2519_2020-16-04   0.346   0.000     0     0     0
#> 46          mai recbut 9.2_r2519_2020-16-04  24.218  14.850   211   300   300
#> 47          mai    amf 9.2_r2519_2020-16-04   0.158   0.000     0     0     0
#> 48          mai recphy 9.2_r2519_2020-16-04  14.492   9.323   204   251   251
#> 49          mai    amf 9.2_r2519_2020-16-04   0.266   0.000     0     0     0
#> 50          mai recphy 9.2_r2519_2020-16-04  13.546   8.585   191   239   239
#> 51          mai    amf 9.2_r2519_2020-16-04   0.286   0.000     0     0     0
#> 52          mai recphy 9.2_r2519_2020-16-04  12.515   8.010   191   235   235
#> 53          mai    amf 9.2_r2519_2020-16-04   0.287   0.000     0     0     0
#> 54          mai recphy 9.2_r2519_2020-16-04  12.174   7.853   199   246   246
#> 55          mai    amf 9.2_r2519_2020-16-04   0.319   0.000     0     0     0
#> 56          mai recphy 9.2_r2519_2020-16-04  16.046   9.447   203   260   260
#> 57          poi    amf 9.2_r2519_2020-16-04   0.033   0.000     0     0     0
#> 58          poi    mat 9.2_r2519_2020-16-04   5.650   2.877   143   180     0
#> 59          pdt    amf 9.2_r2519_2020-16-04   0.015   0.000     0     0     0
#> 60          pdt recbut 9.2_r2519_2020-16-04  18.008  15.307   158   250   250
#> 61          riz    amf 9.2_r2519_2020-16-04   1.870   0.000     0     0     0
#> 62          riz recphy 9.2_r2519_2020-16-04   9.479   3.432   212   255   255
#> 63          col    amf 9.2_r2519_2020-16-04   0.153   0.000     0     0     0
#> 64          col    mat 9.2_r2519_2020-16-04  12.789   5.708   494   550     0
#> 65          sor    amf 9.2_r2519_2020-16-04   0.953   0.000     0     0     0
#> 66          sor recphy 9.2_r2519_2020-16-04   8.930   6.192   215   252   252
#> 67          soj    amf 9.2_r2519_2020-16-04   0.087   0.000     0     0     0
#> 68          soj    mat 9.2_r2519_2020-16-04   6.554   2.899   231   263     0
#> 69          fra    amf 9.2_r2519_2020-16-04   0.925   0.000    61     0     0
#> 70          fra recphy 9.2_r2519_2020-16-04   3.091   1.952    61   133   133
#> 71          bet    amf 9.2_r2519_2020-16-04   0.582   0.219     0     0     0
#> 72          bet recbut 9.2_r2519_2020-16-04  31.268  25.226   254   335   335
#> 73          tou    amf 9.2_r2519_2020-16-04   0.546   0.000     0     0     0
#> 74          tou recphy 9.2_r2519_2020-16-04  12.047   4.546   198   232   232
#> 75          fou    amf 9.2_r2519_2020-16-04   0.691   0.691     0     0     0
#> 76          fou   date 9.2_r2519_2020-16-04   4.646   4.476   153    90    90
#> 77          fou    amf 9.2_r2519_2020-16-04   0.971   0.668   153    90    90
#> 78          fou   date 9.2_r2519_2020-16-04   2.913   2.026   153    90    90
#> 79          fou    amf 9.2_r2519_2020-16-04   0.806   0.454   153    90    90
#> 80          fou    mat 9.2_r2519_2020-16-04   2.624   1.789   153   264    90
#> 81          tom    amf 9.2_r2519_2020-16-04   0.636   0.000     0     0     0
#> 82          tom    mat 9.2_r2519_2020-16-04   3.798   2.106   356   417     0
#> 83          vig    amf 9.2_r2519_2020-16-04   0.029   0.000     0     0     0
#> 84          vig    mat 9.2_r2519_2020-16-04   2.162   0.000   523   640   616
#> 85          ble    amf 9.2_r2519_2020-16-04   0.821   0.000     0     0     0
#> 86          ble    mat 9.2_r2519_2020-16-04  16.698   7.504   514   556     0
#>      laimax  cprecip      cet QNplante    Qles
#> 1     1.124  293.000   70.863   48.365  44.226
#> 2     4.109  559.200  409.053  227.889  49.562
#> 3     0.991  117.000   54.237   27.909   0.123
#> 4     3.027  411.000  366.694  184.703   0.204
#> 5     0.968  413.000   86.123   32.368   1.611
#> 6     3.147  575.500  373.552  169.336   1.611
#> 7     1.035  187.000   64.466   29.772   5.043
#> 8     3.157  418.000  376.614  173.291   5.118
#> 9     1.139  373.500   72.397   35.083   2.936
#> 10    3.118  619.000  402.180  179.992   2.986
#> 11    0.689  586.365  251.707   47.052  16.059
#> 12    2.651  749.065  581.409  175.058  16.059
#> 13    0.066   76.000   78.429    2.136   0.000
#> 14    7.453 2856.000 1235.603  675.971   0.000
#> 15    0.413  233.000  131.691   10.075 169.361
#> 16    3.180 1118.000  767.356   79.212 362.551
#> 17    0.216  243.903  237.598    1.422   0.000
#> 18    2.640 1061.702  678.398  165.985 249.523
#> 19    1.048   76.000   90.219   61.374   0.390
#> 20    2.981  214.000  256.816  108.665   0.740
#> 21 1240.745  194.485  144.910   31.630   0.000
#> 22 1240.912  259.487  149.823   33.559   0.833
#> 23    3.691  194.500  129.519   91.976   0.000
#> 24    5.174  412.500  188.726  118.066  12.982
#> 25    1.902  100.500  108.344   23.417   0.000
#> 26    1.918  411.922  194.028   90.218   1.988
#> 27    0.477   75.000   43.187    6.324   0.000
#> 28    2.319  131.736  127.128   29.473   0.000
#> 29    0.273   74.500   44.918    4.438   0.000
#> 30    0.909  128.981  147.062  106.334   0.000
#> 31    1.088  163.500  114.404   66.818   0.000
#> 32    4.870  424.500  428.653  190.862   0.000
#> 33    0.410   68.500   80.975    5.100   0.000
#> 34    2.415  462.961  424.615   62.875   0.000
#> 35    0.800    0.000    6.100   32.014   0.000
#> 36    4.824   38.400  151.601  104.012   0.000
#> 37    1.451   87.500   93.371   27.543   0.377
#> 38    5.956  343.500  370.582   77.117   0.377
#> 39    0.471  139.000   32.600   68.692   4.266
#> 40    1.990  314.093  295.151  107.884   5.217
#> 41    1.990  314.093  298.293    0.459   5.217
#> 42    1.990  735.630  708.425   53.566   5.217
#> 43    0.000   70.000   19.551    1.291  45.868
#> 44    0.000   86.500   60.659   16.100  45.868
#> 45    0.528  110.000   54.766   11.393   1.449
#> 46    5.800  668.500  627.275  321.271   2.778
#> 47    0.307  184.000   92.497    3.794   0.509
#> 48    3.982  500.500  482.760  177.436   0.571
#> 49    0.415   74.000   65.965    7.341   1.026
#> 50    3.784  459.000  473.195  162.923   1.026
#> 51    0.497   68.500   72.257    8.076   1.263
#> 52    3.640  388.000  451.393  167.745   1.263
#> 53    0.448   72.000   71.400    9.797   0.172
#> 54    3.872  338.000  449.629  159.248   0.172
#> 55    0.536  103.000   76.095    9.264   2.218
#> 56    4.067  413.000  494.412  192.872   2.218
#> 57    0.221   51.600   57.756    1.052   0.000
#> 58    2.337  217.200  254.539  134.513   0.000
#> 59    0.035  108.500   52.282    1.119   0.040
#> 60    3.636  408.100  446.499  224.601   0.152
#> 61    1.338  428.300  195.925   76.055   0.351
#> 62    4.633  783.800  546.651  122.520   0.495
#> 63    0.191  110.900   60.849    9.703  21.583
#> 64    3.533  658.596  490.567  189.939  65.143
#> 65    1.442  124.000   77.420   39.126   1.315
#> 66    6.034  206.500  290.520  161.339   1.315
#> 67    0.130   71.000   73.222    4.997   0.000
#> 68    4.523  152.069  343.174   92.942   0.000
#> 69    0.214   20.000   11.245   48.731   0.042
#> 70    0.519   82.698   47.066   63.499   0.042
#> 71    0.574   98.500  108.474   34.034   5.100
#> 72    5.367  474.149  602.842  236.601   5.100
#> 73    0.554  156.500  103.623   17.189   0.001
#> 74    3.375  370.500  435.968  156.673   0.001
#> 75    1.210   48.400   54.436   42.351   0.037
#> 76    4.560  180.025  163.064  105.433   0.178
#> 77    4.560  196.692  182.669   28.454   0.178
#> 78    4.560  352.661  289.534   92.862   0.369
#> 79    4.560  353.523  304.711   38.886   0.369
#> 80    4.560  498.481  414.622   61.890   0.369
#> 81    0.149   59.265   67.316    5.299   0.000
#> 82    0.534  393.330  394.860  120.420   0.002
#> 83    0.024  513.600  196.661   21.383  27.396
#> 84    0.963  726.000  421.707   45.191  28.211
#> 85    0.987  303.000  104.504   40.375   0.033
#> 86    4.765  619.200  396.015  212.760   0.104

get_report_results(workspace = path, file_name = "mod_rapportA.sti")
#>                  P_usm    wlieu ansemis P_iwater ancours ifin P_ichsl group
#> 1 intercrop_pea_barley readingj    2004       70    2004  149       1     0
#> 2 intercrop_pea_barley readingj    2004       70    2004  208       1     0
#>   P_codeplante stade           nomversion masec_n mafruit iflos imats irecs
#> 1          esc   amf 9.2_r2519_2020-16-04   1.580   0.000     0     0     0
#> 2          esc   mat 9.2_r2519_2020-16-04   4.763   2.089   167   208     0
#>   laimax cprecip     cet QNplante Qles
#> 1  0.694   107.0 138.555   31.432    0
#> 2  0.838   163.2 200.642   41.043    0