Getting existing xml files path list per usm from an usms.xml file

Getting existing xml files path list per usm from an usms.xml file

get_usms_files(
  workspace,
  usms_list = NULL,
  usms_file = "usms.xml",
  file_type = NULL,
  javastics = NULL,
  df_output = FALSE,
  workspace_path = lifecycle::deprecated(),
  file_name = lifecycle::deprecated(),
  javastics_path = lifecycle::deprecated()
)

Arguments

workspace

Path of a JavaSTICS workspace (i.e. containing the STICS XML input files)

usms_list

Vector of usms names (Optional)

usms_file

Path (including name) of a USM XML file.

file_type

Vector of file(s) type to get (if not given, all types are returned, see details)

javastics

Path of JavaSTICS Optional, only needed if the plant files are not in the workspace (in this case the plant files used are those included in the JavaSTICS distribution)

df_output

logical if TRUE returning a data.frame, otherwise returning a named list if FALSE (default)

workspace_path

workspace_path is no longer supported, use workspace instead.

file_name

file_name is no longer supported, use usms_file instead.

javastics_path

javastics_path is no longer supported, use javastics instead.

Value

A named list with existing files path in each usm element

Details

The possible values for file_type are: "fplt", "finit", "fclim1", "fclim2", "fstation", "ftec", "sols", "pargen" and "parnew"

See also

Examples

if (FALSE) {

get_usms_files(
  workspace = "/path/to/workspace",
  javastics = "/path/to/JavaSTICS/folder"
)

get_usms_files(
  workspace = "/path/to/workspace",
  javastics = "/path/to/JavaSTICS/folder",
  usm_list = c("usm1", "usm3")
)

get_usms_files(
  workspace = "/path/to/workspace",
  file_type = c("finit", "ftec")
)
}