Set (replace) STICS input file parameters

Replace or set an input parameter from a pre-existing STICS input file.

set_param_txt(
  dirpath = getwd(),
  param,
  value,
  add = FALSE,
  plant = 1,
  variety = NULL,
  layer = NULL
)

set_usm_txt(filepath = "new_travail.usm", param, value, add = F)

set_station_txt(filepath = "station.txt", param, value, add = F)

set_ini_txt(filepath = "ficini.txt", param, value, add = F)

set_general_txt(filepath = "tempopar.sti", param, value, add = F)

set_tmp_txt(filepath = "tempoparv6.sti", param, value, add = F)

set_plant_txt(filepath = "ficplt1.txt", param, value, add = F, variety = NULL)

set_tec_txt(filepath = "fictec1.txt", param, value, add = F)

set_soil_txt(filepath = "param.sol", param, value, layer = NULL)

Arguments

dirpath
USM directory path

param
Parameter name

value
New parameter value

add
Boolean. Append input to existing file (add to the list)

plant
Plant index. Optional, only for plant or technical parameters

variety
Integer. The plant variety to set the parameter value.

layer
The soil layer if any (only concerns soil-related parameters)

filepath
Path to the parameter file

Details

The plant parameter can be either equal to 1, 2 for the associated plant in the case of intercrop, or c(1,2) for both Principal and associated plants. get_var_info is a helper function that returns all possible output variables.

Note

gen_varmod is not used by set_param_txt. To replace the output variables required from STICS, please directly call gen_varmod.

Examples

if (FALSE) {

# Getting example data path
path = get_examples_path( file_type = "txt" )

# Replace the interrow distance parameter to 0.01:
set_param_txt(dirpath = path, param = "interrang", value = 0.01)

# Change the value of durvieF for the current variety:
set_param_txt(dirpath = path, param = "durvieF", value = 245)

# Change the value of durvieF for another variety:
set_param_txt(dirpath = path, param = "durvieF", variety = "Nefer", value = 178)

# Change the value of infil for a given layer:
set_param_txt(dirpath = path, param = "infil", layer = 2, value = 60)

# If the parameter is found in several files, use the set_* functions direclty, e.g.
# cailloux is found in the general file ("codetycailloux") and the soil file. If we want to
# change its value in the soil file, we use set_soil_txt():
set_soil_txt(filepath = file.path(path, "param.sol"), param = "cailloux", layer = 2, value = 1)
}