Generates files to force parameters values in Stics simulations

Generates a param.sti file and sets code optim in new_travail.usm to force parameters values in Stics simulations (this function is typically called before SticsOnR::run_stics())

force_param_values(
  workspace,
  values,
  javastics,
  param_values = lifecycle::deprecated()
)

Arguments

workspace

Path of the workspace containing the Stics (txt) input files.

values

named vector of parameter values to force. See Details for more information.

javastics

Path of JavaStics.

param_values

old_param old_param is no longer supported, use new_param instead.

Value

A logical status TRUE if successfull, FALSE otherwise

Details

This function operates on Stics text input files. Do not use it before calling gen_usms_xml2txt(), otherwise param.sti and new_travail.usm files will be overwritten.

This function has been created to be called before SticsOnR::run_stics(). It can not work with SticsOnR::run_javastics(), that will overwrite param.sti and new_travail.usm files.

param_values can contain NA. In this case, the corresponding parameter(s) will not be forced (default value(s), i.e. read in Stics input files, will be used). If param_values==NA or param_values==NULL, not any parameter will be forced (all default values used).

See also

Examples

if (FALSE) {
example_txt_dir <- get_examples_path(file_type = "txt")
force_param_values(example_txt_dir,
    setNames(object = c(220, 330), c("stlevamf", "stamflax")),
    javastics = "/path/to/JavaSTICS/folder")
}