Generate Stics tec xml file(s) from a template or an input file

Generate Stics tec xml file(s) from a template or an input file

gen_tec_xml(
  param_table = NULL,
  tec_in_file = NULL,
  out_path = getwd(),
  stics_version = "last"
)

Arguments

param_table
a table (df, tibble) containing parameters to use (see details)

tec_in_file
file path to an XML file (optional, if not povided, uses a template from the package corresponding to stics_version)

out_path
path to an optional folder where to write the output file(s)

stics_version
the stics version to use (optional, default to last). Only used if tec_in_file= NULL, see details.

Value

an invisible xmlDocument object or a list of

Details

Please see get_stics_versions_compat() for the full list of stics versions that can be used for the argument stics_version. param_table is a data.frame with the following format:

Tec_name	julres_1	coderes_1	qres_1	Crespc_1	CsurNres_1

USM_2017_T1_CI_tec.xml	999	1	9	42	90

BIN_CANPC_05_SEC_220-0-0_34K_CANPC05T3_Q_tec.xml	110	1	9	42	90

BIN_AGT_04_IRR_220-0-0_33K_AGT04T2_Q_tec.xml	73	1	9	42	90

AGA_ARB_13_IRR_220-0-0_37K_ARB13_C_tec.xml	82	1	9	42	90

AGA_ARB_13_SEC_220-0-0_37K_ARB13_C_tec.xml	82	1	9	42	90

FRA_ARB_11_SEC_220-0-0_38K_E_tec.xml	70	1	9	42	90

MAG_ARB_09_SEC_220-0-0_38K_E_tec.xml	81	1	9	42	90

MAG_ARV_12_IRR_220-0-0_36K_ARV12_C_tec.xml	100	1	9	42	90

MAG_ARV_12_SEC_220-0-0_36K_ARV12_C_tec.xml	100	1	9	42	90

FRA_ARB_12_SEC_220-0-0_31K_ARB12_C_tec.xml	92	1	9	42	90

FRA_ARB_13_SEC_220-0-0_37K_ARB13_C_tec.xml	82	1	9	42	90

The first column gives the tec file name (to be generated), all following columns give the parameter value to put in the file, and each line denotes a separate tec file (for e.g. several USMs).

The first column name must contain the keyword tec or Tec or TEC as a prefix to be detected (as shown in the table extract above).

Examples

if (FALSE) {

xl_path <-  download_usm_xl(xl_name = "inputs_stics_example.xlsx")
tec_param_df <- read_params_table(file_path = xl_path, sheet_name = "Tec")
gen_tec_xml(out_path = "/path/to/dest/dir", param_table = tec_param_df)

}