Generating one or several usms directories from a javastics workspace content

The function creates sets of input files for one or multiple usms from usms data stored in a JavaStics workspace. For multiple usms, sets will be generated into individual folders named with usm names.. Observations files will be also copied if they are named [usm_name].obs For one usm, files will be generated either in the workspace directory or in a subdirectory.

gen_usms_xml2txt(
  javastics_path,
  workspace_path = NULL,
  target_path = NULL,
  usms_list = c(),
  verbose = TRUE,
  dir_per_usm_flag = TRUE
)

Arguments

javastics_path
Path of JavaStics installation directory

workspace_path
Path of a JavaStics workspace (Optional), if not provided the current workspace stored in JavaStics preferences will be used.

target_path
The path of the directory where to create usms directories (Optional), if not provided the JavaStics workspace will be used as root

usms_list
List of usms to generate (Optional). If not provided, all usms contained in workspace_path/usms.xml file will be generated.

verbose
Logical value for displaying information while running (TRUE) ot not (FALSE)

dir_per_usm_flag
logical, TRUE if one want to create one directory per USM, FALSE if USM files are generated in the target_path (only useful for usms_list of size one)

Value

A list with named elements: usms_path : created directories paths (for storing Stics input files), files : generated files list (in JavaStics workspace origin), copy_status : logical value vector, indicating if all files have been generated for each usm obs_copy_status : logical value vector, indicating if observaion files have been successfully copied in usms directories

Examples

if (FALSE) {
javastics <- "/path/to/javastics"
javastics_workspace <- "/path/to/workspace"

# For all usms
gen_usms_xml2txt(javastics, javastics_workspace)

# For an usms list
usm_list <- c("usm1", "usm2")
gen_usms_xml2txt(javastics, javastics_workspace, usm_list)

# For one usm
gen_usms_xml2txt(javastics, javastics_workspace, dir_per_usm_flag=F, "usm1")

}