Read STICS input parameters from text files

Read STICS model input parameters from a usm in text format (STICS input) Generally used after calling building a usm with JavaSTICS.

Read a specific STICS model input parameter file. Users would generally use the wrapper get_param_txt() instead.

get_param_txt(
  workspace,
  param = NULL,
  plant_id = NULL,
  variety = NULL,
  value_id = NULL,
  exact = FALSE,
  stics_version = "latest",
  dirpath = lifecycle::deprecated(),
  ...
)

get_ini_txt(
  file = "ficini.txt",
  stics_version,
  filepath = lifecycle::deprecated()
)

get_general_txt(file = "tempopar.sti", filepath = lifecycle::deprecated())

get_tmp_txt(file = "tempoparv6.sti", filepath = lifecycle::deprecated())

get_plant_txt(
  file = "ficplt1.txt",
  variety = NULL,
  filepath = lifecycle::deprecated()
)

get_tec_txt(
  file = "fictec1.txt",
  stics_version = "latest",
  several_fert = NULL,
  several_thin = NULL,
  is_pasture = NULL,
  filepath = lifecycle::deprecated(),
  ...
)

get_soil_txt(
  file = "param.sol",
  stics_version,
  filepath = lifecycle::deprecated()
)

get_station_txt(file = "station.txt", filepath = lifecycle::deprecated())

get_usm_txt(
  file = "new_travail.usm",
  plant_id = NULL,
  filepath = lifecycle::deprecated()
)

Arguments

workspace

Path of the workspace containing the STICS (txt) input files.

param

Vector of parameter names. Optional, if not provided, the function returns an object with all parameters.

plant_id

plant index (1, 2), default(NULL) calculated from from plant number in STICS initialization file

variety

Integer. The plant variety to get the parameter from.

value_id

index of technical interventions to be used to retrieve parameter values, or layer index for soil parameters

exact

Boolean indicating if the function must return results only for exact match.

stics_version

An optional version name as listed in get_stics_versions_compat() return

dirpath

dirpath is no longer supported, use workspace instead.

...

Further arguments to pass (for future-proofing only)

file

File path

filepath

filepath is no longer supported, use file instead.

several_fert

Is there several fertilization in the USM ? See details.

several_thin

Is there several thinning in the USM ? See details.

is_pasture

Is the plant a pasture ? See details.

Value

A list of parameters value(s), or if param = NULL a list of all parameters:

ini

Initialization parameters

general

General parameters

tec

Technical parameters

plant

Plant parameters

soil

Soil parameters

station

Station parameters

A list of parameters, depending on the file/function:

ini

Initialization parameters

general

General parameters

tec

Technical parameters

plant

Plant parameters

soil

Soil parameters

station

Station parameters

tmp

Temporary parameters

Details

If the variety is not given and a param is asked, the function will return the values for the variety that is simulated in the USM by checking the variete parameter in the technical file. If param is not provided by the user, the values from all varieties will be returned unless the user ask for a given variety.

several_fert, several_thin and is_pasture are read from the tmp file (tempoparv6.sti). get_param_txt() does it automatically. If you absolutely need to use directly get_tec_txt, please see example.

Note

Users would generally use get_param_txt to identify parameters names and values and pass them to other functions.

The functions are compatible with intercrops. Users generally only use get_param_txt(), which is a wrapper for all these functions.

See also

Examples

path <- get_examples_path(file_type = "txt")

# Getting the interrow distance parameter value
get_param_txt(path, param = "interrang")
#> $tec
#> $tec$plant1
#> $tec$plant1$interrang
#> [1] 2
#> 
#> 
#> 

# Getting varietal parameters values
# Get the leaf lifespan of the variety used in the usm:
get_param_txt(workspace = path, param = "durvieF")
#> $plant
#> $plant$plant1
#> $plant$plant1$durvieF
#> Furio 
#>   200 
#> 
#> 
#> 
# Get the leaf lifespan of another variety available in the plant file:
get_param_txt(workspace = path, param = "durvieF", variety = "Furio")
#> $plant
#> $plant$plant1
#> $plant$plant1$durvieF
#> Furio 
#>   200 
#> 
#> 
#> 
# To get the values for several (or all) varieties, either put all varieties:
varieties <- c("Pactol", "Cherif", "Furio", "Dunia", "Volga", "Cecilia")
get_param_txt(workspace = path, param = "durvieF", variety = varieties)
#> $plant
#> $plant$plant1
#> $plant$plant1$durvieF
#>  Pactol  Cherif   Furio   Dunia   Volga Cecilia 
#>     200     200     200     200     200     200 
#> 
#> 
#> 
# Or get it from the output of the function returning all parameters:
get_param_txt(workspace = path)$plant$plant1$durvieF
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        200        200        200        200        200        200        200 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>        200        200        200        200        200        200        200 
#>    magrite    clarica 
#>        200        200 

# Get parameters for a specific plant
get_param_txt(workspace = path, plant_id = 1)
#> $usm
#> $usm$codesimul
#> [1] "culture"
#> 
#> $usm$codoptim
#> [1] 0
#> 
#> $usm$codesuite
#> [1] 0
#> 
#> $usm$nbplantes
#> [1] 1
#> 
#> $usm$nom
#> [1] "maize"
#> 
#> $usm$datedebut
#> [1] 112
#> 
#> $usm$datefin
#> [1] 360
#> 
#> $usm$finit
#> [1] "mais_ini.xml"
#> 
#> $usm$numsol
#> [1] 1
#> 
#> $usm$nomsol
#> [1] "solmais"
#> 
#> $usm$fstation
#> [1] "climaisj_sta.xml"
#> 
#> $usm$fclim1
#> [1] "climaisj.1996"
#> 
#> $usm$fclim2
#> [1] "climaisj.1996"
#> 
#> $usm$nbans
#> [1] 1
#> 
#> $usm$culturean
#> [1] 1
#> 
#> $usm$fplt1
#> [1] "corn_plt.xml"
#> 
#> $usm$ftec1
#> [1] "Mais_tec.xml"
#> 
#> $usm$flai1
#> [1] "null"
#> 
#> 
#> $ini
#> $ini$nbplantes
#> [1] 1
#> 
#> $ini$plant
#> $ini$plant$plant1
#> $ini$plant$plant1$stade0
#> [1] "snu"
#> 
#> $ini$plant$plant1$lai0
#> [1] 0
#> 
#> $ini$plant$plant1$magrain0
#> [1] 0
#> 
#> $ini$plant$plant1$zrac0
#> [1] 0
#> 
#> $ini$plant$plant1$code_acti_reserve
#> [1] 2
#> 
#> $ini$plant$plant1$maperenne0
#> [1] 0
#> 
#> $ini$plant$plant1$QNperenne0
#> [1] 0
#> 
#> $ini$plant$plant1$masecnp0
#> [1] 0
#> 
#> $ini$plant$plant1$QNplantenp0
#> [1] 0
#> 
#> $ini$plant$plant1$masec0
#> [1] 0
#> 
#> $ini$plant$plant1$QNplante0
#> [1] 0
#> 
#> $ini$plant$plant1$restemp0
#> [1] 0
#> 
#> $ini$plant$plant1$densinitial
#> [1] 0 0 0 0 0
#> 
#> 
#> $ini$plant$plant2
#> $ini$plant$plant2$stade0
#> [1] ""
#> 
#> $ini$plant$plant2$lai0
#> [1] ""
#> 
#> $ini$plant$plant2$magrain0
#> [1] ""
#> 
#> $ini$plant$plant2$zrac0
#> [1] ""
#> 
#> $ini$plant$plant2$code_acti_reserve
#> [1] 2
#> 
#> $ini$plant$plant2$maperenne0
#> [1] 0
#> 
#> $ini$plant$plant2$QNperenne0
#> [1] 0
#> 
#> $ini$plant$plant2$masecnp0
#> [1] 0
#> 
#> $ini$plant$plant2$QNplantenp0
#> [1] 0
#> 
#> $ini$plant$plant2$masec0
#> [1] ""
#> 
#> $ini$plant$plant2$QNplante0
#> [1] ""
#> 
#> $ini$plant$plant2$restemp0
#> [1] ""
#> 
#> $ini$plant$plant2$densinitial
#> [1] "     "
#> 
#> 
#> 
#> $ini$Hinitf
#> [1] 23.5 21.6 23.9 27.6  0.0
#> 
#> $ini$NO3initf
#> [1] 32 12  9  0  0
#> 
#> $ini$NH4initf
#> [1] 0 0 0 0 0
#> 
#> $ini$Sdepth0
#> [1] 0
#> 
#> $ini$Sdry0
#> [1] 0
#> 
#> $ini$Swet0
#> [1] 0
#> 
#> $ini$ps0
#> [1] 0
#> 
#> 
#> $general
#> $general$nbresidus
#> [1] 21
#> 
#> $general$codeinnact
#> [1] 1
#> 
#> $general$codeh2oact
#> [1] 1
#> 
#> $general$codeminopt
#> [1] 2
#> 
#> $general$iniprofil
#> [1] 2
#> 
#> $general$codeprofmes
#> [1] 1
#> 
#> $general$codeinitprec
#> [1] 1
#> 
#> $general$codemsfinal
#> [1] 1
#> 
#> $general$codeactimulch
#> [1] 1
#> 
#> $general$codefrmur
#> [1] 1
#> 
#> $general$codemicheur
#> [1] 1
#> 
#> $general$codeoutscient
#> [1] 2
#> 
#> $general$codeseprapport
#> [1] 2
#> 
#> $general$separateurrapport
#> [1] ";"
#> 
#> $general$codesensibilite
#> [1] 2
#> 
#> $general$codesnow
#> [1] 2
#> 
#> $general$flagecriture
#> [1] 31
#> 
#> $general$parsurrg
#> [1] 0.48
#> 
#> $general$coefb
#> [1] 0.0815
#> 
#> $general$proprac
#> [1] 0.2
#> 
#> $general$y0msrac
#> [1] 0.7
#> 
#> $general$dacohes
#> [1] 0.7
#> 
#> $general$daseuilbas
#> [1] 1.4
#> 
#> $general$daseuilhaut
#> [1] 2
#> 
#> $general$beta
#> [1] 1.4
#> 
#> $general$lvopt
#> [1] 0.5
#> 
#> $general$difN
#> [1] 0.045
#> 
#> $general$plNmin
#> [1] 10
#> 
#> $general$irrlev
#> [1] 20
#> 
#> $general$QNpltminINN
#> [1] 0
#> 
#> $general$codesymbiose
#> [1] 2
#> 
#> $general$codefxn
#> [1] 2
#> 
#> $general$tmin_mineralisation
#> [1] 0
#> 
#> $general$FTEMh
#> [1] 0.12
#> 
#> $general$FTEMha
#> [1] 25
#> 
#> $general$TREFh
#> [1] 15
#> 
#> $general$FTEMr
#> [1] 0.103
#> 
#> $general$FTEMra
#> [1] 12
#> 
#> $general$TREFr
#> [1] 15
#> 
#> $general$GMIN1
#> [1] 7e-04
#> 
#> $general$GMIN2
#> [1] 0.02519
#> 
#> $general$GMIN3
#> [1] 0.015
#> 
#> $general$GMIN4
#> [1] 0.112
#> 
#> $general$GMIN5
#> [1] 8.5
#> 
#> $general$GMIN6
#> [1] 0.06
#> 
#> $general$GMIN7
#> [1] 11
#> 
#> $general$Wh
#> [1] 0.105
#> 
#> $general$pHminvol
#> [1] 5.5
#> 
#> $general$pHmaxvol
#> [1] 8.5
#> 
#> $general$Vabs2
#> [1] 2
#> 
#> $general$Xorgmax
#> [1] 400
#> 
#> $general$hminm
#> [1] 0.25
#> 
#> $general$hoptm
#> [1] 0.9
#> 
#> $general$alphaph
#> [1] 0.005
#> 
#> $general$dphvolmax
#> [1] 1
#> 
#> $general$phvols
#> [1] 8.6
#> 
#> $general$fhminsat
#> [1] 0.5
#> 
#> $general$fredkN
#> [1] 0.25
#> 
#> $general$fredlN
#> [1] 0.5
#> 
#> $general$fNCbiomin
#> [1] 0.04
#> 
#> $general$fredNsup
#> [1] 0.5
#> 
#> $general$Primingmax
#> [1] 3
#> 
#> $general$hminn
#> [1] 0.3
#> 
#> $general$hoptn
#> [1] 0.9
#> 
#> $general$pHminnit
#> [1] 4
#> 
#> $general$pHmaxnit
#> [1] 7.2
#> 
#> $general$nh4_min
#> [1] 1
#> 
#> $general$pHminden
#> [1] 5.6
#> 
#> $general$pHmaxden
#> [1] 9.2
#> 
#> $general$wfpsc
#> [1] 0.62
#> 
#> $general$tdenitopt_gauss
#> [1] 47
#> 
#> $general$scale_tdenitopt
#> [1] 25
#> 
#> $general$Kd
#> [1] 148
#> 
#> $general$kdesat
#> [1] 3
#> 
#> $general$code_vnit
#> [1] 1
#> 
#> $general$fnx
#> [1] 0.8
#> 
#> $general$vnitmax
#> [1] 27.3
#> 
#> $general$Kamm
#> [1] 24
#> 
#> $general$code_tnit
#> [1] 2
#> 
#> $general$tnitmin
#> [1] 5
#> 
#> $general$tnitopt
#> [1] 30
#> 
#> $general$tnitopt2
#> [1] 35
#> 
#> $general$tnitmax
#> [1] 58
#> 
#> $general$tnitopt_gauss
#> [1] 32.5
#> 
#> $general$scale_tnitopt
#> [1] 16
#> 
#> $general$code_rationit
#> [1] 2
#> 
#> $general$rationit
#> [1] 0.0016
#> 
#> $general$code_hourly_wfps_nit
#> [1] 1
#> 
#> $general$code_pdenit
#> [1] 1
#> 
#> $general$cmin_pdenit
#> [1] 1
#> 
#> $general$cmax_pdenit
#> [1] 6
#> 
#> $general$min_pdenit
#> [1] 1
#> 
#> $general$max_pdenit
#> [1] 20
#> 
#> $general$code_ratiodenit
#> [1] 2
#> 
#> $general$ratiodenit
#> [1] 0.2
#> 
#> $general$code_hourly_wfps_denit
#> [1] 1
#> 
#> $general$pminruis
#> [1] 5
#> 
#> $general$diftherm
#> [1] 0.00537
#> 
#> $general$bformnappe
#> [1] 0.77
#> 
#> $general$rdrain
#> [1] 1
#> 
#> $general$psihumin
#> [1] -1.5
#> 
#> $general$psihucc
#> [1] -0.03
#> 
#> $general$prophumtasssem
#> [1] 1.2
#> 
#> $general$prophumtassrec
#> [1] 1
#> 
#> $general$codhnappe
#> [1] 1
#> 
#> $general$distdrain
#> [1] 1000
#> 
#> $general$proflabour
#> [1] 30
#> 
#> $general$proftravmin
#> [1] 5
#> 
#> $general$codetycailloux
#> [1] 2
#> 
#> $general$masvolcx
#>  [1] 2.20 1.80 2.10 2.30 2.50 2.65 2.30 2.20 1.50 0.00
#> 
#> $general$hcccx
#>  [1]  7 16 11  5  3  1  5  5 26  0
#> 
#> $general$codetypeng
#> [1] 1
#> 
#> $general$engamm
#> [1] 0.50 0.75 1.00 1.00 1.00 1.00 0.00 0.50
#> 
#> $general$orgeng
#> [1] 46.0 42.0 38.0 38.0 38.0 38.0 25.0  0.2
#> 
#> $general$deneng
#> [1] 0.15 0.13 0.10 0.10 0.10 0.10 0.20 0.05
#> 
#> $general$voleng
#> [1] 0.15 0.30 0.35 0.35 0.25 0.25 0.00 0.05
#> 
#> $general$codetypres
#> [1] 21
#> 
#> $general$CroCo
#>  [1] 1 1 1 1 1 1 1 1 0 0 1 1 1 1 1 1 1 1 0 1 1
#> 
#> $general$akres
#>  [1] 0.000 0.000 0.064 0.080 0.064 0.700 0.140 0.000 0.000 0.000 0.098 0.098
#> [13] 0.064 0.080 0.064 0.700 0.140 0.098 0.000 0.030 0.030
#> 
#> $general$bkres
#>  [1]  0.60  0.60 -0.26 -0.48  0.05  0.00  0.00  0.60  0.00  0.00  0.76  0.76
#> [13] -0.26 -0.48  0.05  0.00  0.00  0.76  0.00  1.13  1.13
#> 
#> $general$awb
#>  [1] 30.1 30.1 28.8 33.0 12.7  5.5  5.5 30.1  0.0  0.0 15.4 15.4 28.8 33.0 12.7
#> [16]  5.5  5.5 15.4  0.0 15.2 15.2
#> 
#> $general$bwb
#>  [1] -275 -275 -228 -400  -58    0    0 -275    0    0  -76  -76 -228 -400  -58
#> [16]    0    0  -76    0  -80  -80
#> 
#> $general$cwb
#>  [1]  6.0  6.0 11.5  9.5  6.2  5.5  5.5  6.0  0.0  0.0  7.8  7.8 11.5  9.5  6.2
#> [16]  5.5  5.5  7.8  0.0  7.8  7.8
#> 
#> $general$ahres
#>  [1]  0.91  0.91 36.50  1.00 10.00  0.18  0.05  0.91  0.00  0.00  0.73  0.73
#> [13] 36.50  1.00 10.00  0.18  0.05  0.73  0.00  0.91  0.91
#> 
#> $general$bhres
#>  [1]   16.2   16.2 1305.0   29.4  329.0    0.0    0.0   16.2    0.0    0.0
#> [11]   10.2   10.2 1305.0   29.4  329.0    0.0    0.0   10.2    0.0   25.9
#> [21]   25.9
#> 
#> $general$kbio
#>  [1] 0.0122 0.0122 0.0025 0.0015 0.0027 0.0110 0.0035 0.0122 0.0000 0.0000
#> [11] 0.0076 0.0076 0.0025 0.0015 0.0027 0.0110 0.0035 0.0076 0.0000 0.0076
#> [21] 0.0076
#> 
#> $general$yres
#>  [1] 0.62 0.62 0.90 0.82 0.62 0.42 0.54 0.62 0.00 0.00 0.62 0.62 0.90 0.82 0.62
#> [16] 0.42 0.54 0.62 0.00 0.62 0.62
#> 
#> $general$CNresmin
#>  [1]  8  8 12 15  6  5  3  8  0  0  8  8 12 15  6  5  3  8  0  8  8
#> 
#> $general$CNresmax
#>  [1] 100 100  27  20  20   8   4 100   0   0 100 100  27  20  20   8   4 100   0
#> [20] 100 100
#> 
#> $general$qmulchruis0
#>  [1] 1 1 1 1 1 1 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $general$mouillabilmulch
#>  [1] 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0
#> [20] 0.0 0.0
#> 
#> $general$kcouvmlch
#>  [1] 0.37 0.37 0.37 0.37 0.37 0.37 0.37 0.05 0.00 0.00 0.00 0.00 0.00 0.00 0.00
#> [16] 0.00 0.00 0.00 0.00 0.00 0.00
#> 
#> $general$albedomulchresidus
#>  [1] 0.10 0.10 0.10 0.10 0.10 0.10 0.10 0.08 0.00 0.00 0.00 0.00 0.00 0.00 0.00
#> [16] 0.00 0.00 0.00 0.00 0.00 0.00
#> 
#> $general$Qmulchdec
#>  [1] 1 1 1 1 1 1 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> 
#> $tec
#> $tec$plant1
#> $tec$plant1$nbjres
#> [1] 1
#> 
#> $tec$plant1$julres
#> [1] 112
#> 
#> $tec$plant1$coderes
#> [1] 1
#> 
#> $tec$plant1$qres
#> [1] 1
#> 
#> $tec$plant1$Crespc
#> [1] 42
#> 
#> $tec$plant1$CsurNres
#> [1] 60
#> 
#> $tec$plant1$Nminres
#> [1] 0
#> 
#> $tec$plant1$eaures
#> [1] 0
#> 
#> $tec$plant1$code_auto_profres
#> [1] 1
#> 
#> $tec$plant1$resk
#> [1] 0.14
#> 
#> $tec$plant1$resz
#> [1] 5
#> 
#> $tec$plant1$nbjtrav
#> [1] 1
#> 
#> $tec$plant1$jultrav
#> [1] 112
#> 
#> $tec$plant1$profres
#> [1] 0
#> 
#> $tec$plant1$proftrav
#> [1] 25
#> 
#> $tec$plant1$iplt0
#> [1] 117
#> 
#> $tec$plant1$profsem
#> [1] 5
#> 
#> $tec$plant1$densitesem
#> [1] 9.5
#> 
#> $tec$plant1$variete
#> [1] 4
#> 
#> $tec$plant1$codetradtec
#> [1] 1
#> 
#> $tec$plant1$interrang
#> [1] 2
#> 
#> $tec$plant1$orientrang
#> [1] 0
#> 
#> $tec$plant1$codedecisemis
#> [1] 2
#> 
#> $tec$plant1$nbjmaxapressemis
#> [1] 30
#> 
#> $tec$plant1$nbjseuiltempref
#> [1] 15
#> 
#> $tec$plant1$nbj_pr_apres_semis
#> [1] 3
#> 
#> $tec$plant1$eau_mini_decisemis
#> [1] 10
#> 
#> $tec$plant1$humirac_decisemis
#> [1] 0.75
#> 
#> $tec$plant1$codestade
#> [1] 2
#> 
#> $tec$plant1$ilev
#> [1] 999
#> 
#> $tec$plant1$iamf
#> [1] 999
#> 
#> $tec$plant1$ilax
#> [1] 999
#> 
#> $tec$plant1$isen
#> [1] 999
#> 
#> $tec$plant1$ilan
#> [1] 999
#> 
#> $tec$plant1$iflo
#> [1] 999
#> 
#> $tec$plant1$idrp
#> [1] 999
#> 
#> $tec$plant1$imat
#> [1] 999
#> 
#> $tec$plant1$irec
#> [1] 999
#> 
#> $tec$plant1$effirr
#> [1] 1
#> 
#> $tec$plant1$codecalirrig
#> [1] 2
#> 
#> $tec$plant1$ratiol
#> [1] 0
#> 
#> $tec$plant1$dosimx
#> [1] 40
#> 
#> $tec$plant1$doseirrigmin
#> [1] 20
#> 
#> $tec$plant1$codedate_irrigauto
#> [1] 3
#> 
#> $tec$plant1$datedeb_irrigauto
#> [1] 0
#> 
#> $tec$plant1$datefin_irrigauto
#> [1] 0
#> 
#> $tec$plant1$stage_start_irrigauto
#> [1] 0
#> 
#> $tec$plant1$stage_end_irrigauto
#> [1] 0
#> 
#> $tec$plant1$codedateappH2O
#> [1] 2
#> 
#> $tec$plant1$nap
#> [1] 16
#> 
#> $tec$plant1$julapI_or_sum_upvt
#>  [1] 178 185 193 198 200 204 207 211 214 218 221 232 239 249 257 264
#> 
#> $tec$plant1$amount
#>  [1] 20 24 29 29 28 31 21 21 23 18 22 16 16 30 29 20
#> 
#> $tec$plant1$codlocirrig
#> [1] 1
#> 
#> $tec$plant1$locirrig
#> [1] 0
#> 
#> $tec$plant1$profmes
#> [1] 120
#> 
#> $tec$plant1$concirr
#> [1] 0.11
#> 
#> $tec$plant1$codedateappN
#> [1] 2
#> 
#> $tec$plant1$codefracappN
#> [1] 1
#> 
#> $tec$plant1$Qtot_N
#> [1] 100
#> 
#> $tec$plant1$napN
#> [1] 1
#> 
#> $tec$plant1$julapN_or_sum_upvt
#> [1] 112
#> 
#> $tec$plant1$`absolute_value/%`
#> [1] 220
#> 
#> $tec$plant1$engrais
#> [1] 2
#> 
#> $tec$plant1$codlocferti
#> [1] 1
#> 
#> $tec$plant1$locferti
#> [1] 0
#> 
#> $tec$plant1$irecbutoir
#> [1] 300
#> 
#> $tec$plant1$ressuite
#> [1] "straw+roots"
#> 
#> $tec$plant1$code_autoressuite
#> [1] 2
#> 
#> $tec$plant1$Stubblevegratio
#> [1] 0
#> 
#> $tec$plant1$codceuille
#> [1] 1
#> 
#> $tec$plant1$nbceuille
#> [1] 1
#> 
#> $tec$plant1$cadencerec
#> [1] 7
#> 
#> $tec$plant1$codrecolte
#> [1] 1
#> 
#> $tec$plant1$codeaumin
#> [1] 2
#> 
#> $tec$plant1$h2ograinmin
#> [1] 0.8
#> 
#> $tec$plant1$h2ograinmax
#> [1] 0.32
#> 
#> $tec$plant1$sucrerec
#> [1] 0.25
#> 
#> $tec$plant1$CNgrainrec
#> [1] 0.02
#> 
#> $tec$plant1$huilerec
#> [1] 0.25
#> 
#> $tec$plant1$coderecolteassoc
#> [1] 2
#> 
#> $tec$plant1$codedecirecolte
#> [1] 2
#> 
#> $tec$plant1$nbjmaxapresrecolte
#> [1] 15
#> 
#> $tec$plant1$codefauche
#> [1] 2
#> 
#> $tec$plant1$code_hautfauche_dyn
#> [1] 2
#> 
#> $tec$plant1$codetempfauche
#> [1] 1
#> 
#> $tec$plant1$codemodfauche
#> [1] 1
#> 
#> $tec$plant1$hautcoupedefaut
#> [1] 0
#> 
#> $tec$plant1$stadecoupedf
#> [1] "rec"
#> 
#> $tec$plant1$nbcoupe
#> [1] 0
#> 
#> $tec$plant1$codepaillage
#> [1] 1
#> 
#> $tec$plant1$couvermulchplastique
#> [1] 0
#> 
#> $tec$plant1$albedomulchplastique
#> [1] 0
#> 
#> $tec$plant1$codrognage
#> [1] 1
#> 
#> $tec$plant1$largrogne
#> [1] 0
#> 
#> $tec$plant1$hautrogne
#> [1] 0
#> 
#> $tec$plant1$biorognem
#> [1] 0
#> 
#> $tec$plant1$codcalrogne
#> [1] 1
#> 
#> $tec$plant1$julrogne
#> [1] 0
#> 
#> $tec$plant1$margerogne
#> [1] 0
#> 
#> $tec$plant1$codeclaircie
#> [1] 1
#> 
#> $tec$plant1$nb_eclair
#> [1] 0
#> 
#> $tec$plant1$codeffeuil
#> [1] 1
#> 
#> $tec$plant1$codhauteff
#> [1] 1
#> 
#> $tec$plant1$codcaleffeuil
#> [1] 1
#> 
#> $tec$plant1$laidebeff
#> [1] 0
#> 
#> $tec$plant1$effeuil
#> [1] 0
#> 
#> $tec$plant1$juleffeuil
#> [1] 0
#> 
#> $tec$plant1$laieffeuil
#> [1] 0
#> 
#> $tec$plant1$codetaille
#> [1] 1
#> 
#> $tec$plant1$jultaille
#> [1] 0
#> 
#> $tec$plant1$codepalissage
#> [1] 1
#> 
#> $tec$plant1$hautmaxtec
#> [1] 0
#> 
#> $tec$plant1$largtec
#> [1] 0.5
#> 
#> $tec$plant1$codabri
#> [1] 1
#> 
#> $tec$plant1$transplastic
#> [1] 0.7
#> 
#> $tec$plant1$surfouvre1
#> [1] 0
#> 
#> $tec$plant1$julouvre2
#> [1] 0
#> 
#> $tec$plant1$surfouvre2
#> [1] 0
#> 
#> $tec$plant1$julouvre3
#> [1] 0
#> 
#> $tec$plant1$surfouvre3
#> [1] 0
#> 
#> $tec$plant1$codejourdes
#> [1] 2
#> 
#> $tec$plant1$juldes
#> [1] 999
#> 
#> $tec$plant1$codeDST
#> [1] 2
#> 
#> $tec$plant1$dachisel
#> [1] 1.1
#> 
#> $tec$plant1$dalabour
#> [1] 1.3
#> 
#> $tec$plant1$rugochisel
#> [1] 0.005
#> 
#> $tec$plant1$rugolabour
#> [1] 0.05
#> 
#> $tec$plant1$codeDSTtass
#> [1] 2
#> 
#> $tec$plant1$profhumsemoir
#> [1] 30
#> 
#> $tec$plant1$dasemis
#> [1] 1.6
#> 
#> $tec$plant1$profhumrecolteuse
#> [1] 30
#> 
#> $tec$plant1$darecolte
#> [1] 1.6
#> 
#> 
#> 
#> $plant
#> $plant$plant1
#> $plant$plant1$codeplante
#> [1] "mai"
#> 
#> $plant$plant1$codemonocot
#> [1] 1
#> 
#> $plant$plant1$alphaco2
#> [1] 1.06
#> 
#> $plant$plant1$tdmin
#> [1] 6
#> 
#> $plant$plant1$tdmax
#> [1] 32
#> 
#> $plant$plant1$codetemp
#> [1] 1
#> 
#> $plant$plant1$codegdh
#> [1] 1
#> 
#> $plant$plant1$coeflevamf
#> [1] 1.44
#> 
#> $plant$plant1$coefamflax
#> [1] 1.07
#> 
#> $plant$plant1$coeflaxsen
#> [1] 1
#> 
#> $plant$plant1$coefsenlan
#> [1] 1.1
#> 
#> $plant$plant1$coeflevdrp
#> [1] 1.13
#> 
#> $plant$plant1$coefdrpmat
#> [1] 1.03
#> 
#> $plant$plant1$coefflodrp
#> [1] 1
#> 
#> $plant$plant1$codephot
#> [1] 2
#> 
#> $plant$plant1$codephot_part
#> [1] 2
#> 
#> $plant$plant1$coderetflo
#> [1] 2
#> 
#> $plant$plant1$stressdev
#> [1] 0.2
#> 
#> $plant$plant1$codebfroid
#> [1] 1
#> 
#> $plant$plant1$jvcmini
#> [1] 7
#> 
#> $plant$plant1$julvernal
#> [1] 274
#> 
#> $plant$plant1$tfroid
#> [1] 6.5
#> 
#> $plant$plant1$ampfroid
#> [1] 10
#> 
#> $plant$plant1$tdmindeb
#> [1] 5
#> 
#> $plant$plant1$tdmaxdeb
#> [1] 25
#> 
#> $plant$plant1$codedormance
#> [1] 3
#> 
#> $plant$plant1$ifindorm
#> [1] 70
#> 
#> $plant$plant1$q10
#> [1] 3
#> 
#> $plant$plant1$idebdorm
#> [1] 300
#> 
#> $plant$plant1$codegdhdeb
#> [1] 2
#> 
#> $plant$plant1$code_WangEngel
#> [1] 2
#> 
#> $plant$plant1$tdoptdeb
#> [1] 11.7
#> 
#> $plant$plant1$tgmin
#> [1] 8
#> 
#> $plant$plant1$codeperenne
#> [1] 1
#> 
#> $plant$plant1$codegermin
#> [1] 1
#> 
#> $plant$plant1$stpltger
#> [1] 35
#> 
#> $plant$plant1$potgermi
#> [1] -1.6
#> 
#> $plant$plant1$nbjgerlim
#> [1] 50
#> 
#> $plant$plant1$propjgermin
#> [1] 1
#> 
#> $plant$plant1$codehypo
#> [1] 1
#> 
#> $plant$plant1$belong
#> [1] 0.022
#> 
#> $plant$plant1$celong
#> [1] 2.04
#> 
#> $plant$plant1$elmax
#> [1] 8
#> 
#> $plant$plant1$nlevlim1
#> [1] 50
#> 
#> $plant$plant1$nlevlim2
#> [1] 50
#> 
#> $plant$plant1$vigueurbat
#> [1] 1
#> 
#> $plant$plant1$laiplantule
#> [1] 0
#> 
#> $plant$plant1$nbfeuilplant
#> [1] 0
#> 
#> $plant$plant1$masecplantule
#> [1] 0.5
#> 
#> $plant$plant1$zracplantule
#> [1] 2
#> 
#> $plant$plant1$phyllotherme
#> [1] 70
#> 
#> $plant$plant1$laicomp
#> [1] 0
#> 
#> $plant$plant1$tcmin
#> [1] 8
#> 
#> $plant$plant1$tcmax
#> [1] 32
#> 
#> $plant$plant1$tcxstop
#> [1] 35
#> 
#> $plant$plant1$codelaitr
#> [1] 1
#> 
#> $plant$plant1$vlaimax
#> [1] 2.2
#> 
#> $plant$plant1$pentlaimax
#> [1] 5.5
#> 
#> $plant$plant1$udlaimax
#> [1] 3
#> 
#> $plant$plant1$ratiodurvieI
#> [1] 1
#> 
#> $plant$plant1$ratiosen
#> [1] 0.8
#> 
#> $plant$plant1$abscission
#> [1] 0
#> 
#> $plant$plant1$parazofmorte
#> [1] 13
#> 
#> $plant$plant1$innturgmin
#> [1] 0.3
#> 
#> $plant$plant1$dlaimin
#> [1] 0
#> 
#> $plant$plant1$codlainet
#> [1] 2
#> 
#> $plant$plant1$tustressmin
#> [1] 0.7
#> 
#> $plant$plant1$durviesupmax
#> [1] 0.4
#> 
#> $plant$plant1$codestrphot
#> [1] 2
#> 
#> $plant$plant1$phobasesen
#> [1] 12
#> 
#> $plant$plant1$dltamsmaxsen
#> [1] 0.01
#> 
#> $plant$plant1$dltamsminsen
#> [1] 0.15
#> 
#> $plant$plant1$alphaphot
#> [1] 200
#> 
#> $plant$plant1$tauxrecouvmax
#> [1] 1
#> 
#> $plant$plant1$tauxrecouvkmax
#> [1] 1
#> 
#> $plant$plant1$pentrecouv
#> [1] 6.3
#> 
#> $plant$plant1$infrecouv
#> [1] 0.85
#> 
#> $plant$plant1$codetransrad
#> [1] 1
#> 
#> $plant$plant1$forme
#> [1] 1
#> 
#> $plant$plant1$rapforme
#> [1] 4
#> 
#> $plant$plant1$adfol
#> [1] 1
#> 
#> $plant$plant1$dfolbas
#> [1] 5
#> 
#> $plant$plant1$dfolhaut
#> [1] 5
#> 
#> $plant$plant1$temax
#> [1] 32
#> 
#> $plant$plant1$teoptbis
#> [1] 32
#> 
#> $plant$plant1$efcroijuv
#> [1] 1.9
#> 
#> $plant$plant1$efcroiveg
#> [1] 3.8
#> 
#> $plant$plant1$efcroirepro
#> [1] 3.8
#> 
#> $plant$plant1$remobres
#> [1] 0.2
#> 
#> $plant$plant1$coefmshaut
#> [1] 0
#> 
#> $plant$plant1$slamin
#> [1] 180
#> 
#> $plant$plant1$envfruit
#> [1] 0.1
#> 
#> $plant$plant1$sea
#> [1] 100
#> 
#> $plant$plant1$code_acti_reserve
#> [1] 2
#> 
#> $plant$plant1$propresP
#> [1] 0.47
#> 
#> $plant$plant1$PropresPN
#> [1] 0.6
#> 
#> $plant$plant1$Efremobil
#> [1] 0.05
#> 
#> $plant$plant1$Propres
#> [1] 0.1
#> 
#> $plant$plant1$tauxmortresP
#> [1] 0.0025
#> 
#> $plant$plant1$Parazoper
#> [1] 25
#> 
#> $plant$plant1$ParazoTmorte
#> [1] 30
#> 
#> $plant$plant1$inilai
#> [1] 0.001
#> 
#> $plant$plant1$resplmax
#> [1] 0.66
#> 
#> $plant$plant1$codemontaison
#> [1] 2
#> 
#> $plant$plant1$codedyntalle
#> [1] 2
#> 
#> $plant$plant1$codetranspitalle
#> [1] 1
#> 
#> $plant$plant1$SurfApex
#> [1] 5e-06
#> 
#> $plant$plant1$SeuilMorTalle
#> [1] 0.02
#> 
#> $plant$plant1$SigmaDisTalle
#> [1] 0.1
#> 
#> $plant$plant1$VitReconsPeupl
#> [1] 0.001
#> 
#> $plant$plant1$SeuilReconsPeupl
#> [1] 800
#> 
#> $plant$plant1$MaxTalle
#> [1] 4000
#> 
#> $plant$plant1$SeuilLAIapex
#> [1] 1
#> 
#> $plant$plant1$tigefeuilcoupe
#> [1] 5
#> 
#> $plant$plant1$codeindetermin
#> [1] 1
#> 
#> $plant$plant1$cgrain
#> [1] 0.05
#> 
#> $plant$plant1$cgrainv0
#> [1] 0.111
#> 
#> $plant$plant1$irazomax
#> [1] 0.566
#> 
#> $plant$plant1$codeir
#> [1] 2
#> 
#> $plant$plant1$irmax
#> [1] 0.53
#> 
#> $plant$plant1$nboite
#> [1] 10
#> 
#> $plant$plant1$allocfrmax
#> [1] 0.6
#> 
#> $plant$plant1$afpf
#> [1] 0.4
#> 
#> $plant$plant1$bfpf
#> [1] 4.32
#> 
#> $plant$plant1$cfpf
#> [1] 0
#> 
#> $plant$plant1$dfpf
#> [1] 0.2
#> 
#> $plant$plant1$spfrmin
#> [1] 0.75
#> 
#> $plant$plant1$spfrmax
#> [1] 1
#> 
#> $plant$plant1$splaimin
#> [1] 0.2
#> 
#> $plant$plant1$splaimax
#> [1] 1
#> 
#> $plant$plant1$codcalinflo
#> [1] 1
#> 
#> $plant$plant1$codetremp
#> [1] 1
#> 
#> $plant$plant1$tminremp
#> [1] 0
#> 
#> $plant$plant1$tmaxremp
#> [1] 40
#> 
#> $plant$plant1$sensanox
#> [1] 0
#> 
#> $plant$plant1$stoprac
#> [1] "lax"
#> 
#> $plant$plant1$sensrsec
#> [1] 0
#> 
#> $plant$plant1$contrdamax
#> [1] 0
#> 
#> $plant$plant1$rayon
#> [1] 0.02
#> 
#> $plant$plant1$codetemprac
#> [1] 2
#> 
#> $plant$plant1$codemortalracine
#> [1] 2
#> 
#> $plant$plant1$coefracoupe
#> [1] 0.5
#> 
#> $plant$plant1$coderacine
#> [1] 2
#> 
#> $plant$plant1$zlabour
#> [1] 25
#> 
#> $plant$plant1$zpente
#> [1] 110
#> 
#> $plant$plant1$zprlim
#> [1] 140
#> 
#> $plant$plant1$draclong
#> [1] 5000
#> 
#> $plant$plant1$debsenrac
#> [1] 1500
#> 
#> $plant$plant1$lvfront
#> [1] 0.05
#> 
#> $plant$plant1$longsperac
#> [1] 11000
#> 
#> $plant$plant1$codedisrac
#> [1] 2
#> 
#> $plant$plant1$kdisrac
#> [1] 0.00158
#> 
#> $plant$plant1$alloperirac
#> [1] 0.23
#> 
#> $plant$plant1$codazorac
#> [1] 2
#> 
#> $plant$plant1$minefnra
#> [1] 0
#> 
#> $plant$plant1$minazorac
#> [1] 0.1
#> 
#> $plant$plant1$maxazorac
#> [1] 1
#> 
#> $plant$plant1$codtrophrac
#> [1] 3
#> 
#> $plant$plant1$repracpermax
#> [1] 0.78
#> 
#> $plant$plant1$repracpermin
#> [1] 0.12
#> 
#> $plant$plant1$krepracperm
#> [1] 1.72
#> 
#> $plant$plant1$repracseumax
#> [1] 0.4
#> 
#> $plant$plant1$repracseumin
#> [1] 0.2
#> 
#> $plant$plant1$krepracseu
#> [1] 1.27
#> 
#> $plant$plant1$code_stress_root
#> [1] 2
#> 
#> $plant$plant1$code_rootdeposition
#> [1] 2
#> 
#> $plant$plant1$parazorac
#> [1] -999
#> 
#> $plant$plant1$code_diff_root
#> [1] 2
#> 
#> $plant$plant1$lvmax
#> [1] -999
#> 
#> $plant$plant1$rapdia
#> [1] -999
#> 
#> $plant$plant1$RTD
#> [1] -999
#> 
#> $plant$plant1$propracfmax
#> [1] -999
#> 
#> $plant$plant1$tletale
#> [1] -5
#> 
#> $plant$plant1$tdebgel
#> [1] 0
#> 
#> $plant$plant1$codgellev
#> [1] 1
#> 
#> $plant$plant1$nbfgellev
#> [1] 3
#> 
#> $plant$plant1$tgellev90
#> [1] -4
#> 
#> $plant$plant1$codgeljuv
#> [1] 1
#> 
#> $plant$plant1$tgeljuv90
#> [1] -4
#> 
#> $plant$plant1$codgelveg
#> [1] 1
#> 
#> $plant$plant1$tgelveg90
#> [1] -4
#> 
#> $plant$plant1$codgelflo
#> [1] 1
#> 
#> $plant$plant1$tgelflo10
#> [1] 0
#> 
#> $plant$plant1$tgelflo90
#> [1] -1
#> 
#> $plant$plant1$h2ofeuilverte
#> [1] 0.9
#> 
#> $plant$plant1$h2ofeuiljaune
#> [1] 0.15
#> 
#> $plant$plant1$h2otigestruc
#> [1] 0.6
#> 
#> $plant$plant1$h2oreserve
#> [1] 0.7
#> 
#> $plant$plant1$h2ofrvert
#> [1] 0.4
#> 
#> $plant$plant1$tempdeshyd
#> [1] 0.005
#> 
#> $plant$plant1$codebeso
#> [1] 1
#> 
#> $plant$plant1$kmax
#> [1] 1.2
#> 
#> $plant$plant1$rsmin
#> [1] 215
#> 
#> $plant$plant1$codeintercept
#> [1] 2
#> 
#> $plant$plant1$mouillabil
#> [1] 0
#> 
#> $plant$plant1$stemflowmax
#> [1] 0
#> 
#> $plant$plant1$kstemflow
#> [1] 0
#> 
#> $plant$plant1$Vmax1
#> [1] 0.0018
#> 
#> $plant$plant1$Kmabs1
#> [1] 50
#> 
#> $plant$plant1$Vmax2
#> [1] 0.017
#> 
#> $plant$plant1$Kmabs2
#> [1] 25000
#> 
#> $plant$plant1$adil
#> [1] 3.5
#> 
#> $plant$plant1$bdil
#> [1] 0.37
#> 
#> $plant$plant1$masecNmax
#> [1] 1
#> 
#> $plant$plant1$INNmin
#> [1] 0.3
#> 
#> $plant$plant1$INNimin
#> [1] -0.5
#> 
#> $plant$plant1$inngrain1
#> [1] 2
#> 
#> $plant$plant1$inngrain2
#> [1] 2
#> 
#> $plant$plant1$bdilmax
#> [1] 0.37
#> 
#> $plant$plant1$codeplisoleN
#> [1] 2
#> 
#> $plant$plant1$adilmax
#> [1] 5.6
#> 
#> $plant$plant1$Nmeta
#> [1] 4.8
#> 
#> $plant$plant1$masecmeta
#> [1] 0.04
#> 
#> $plant$plant1$Nreserve
#> [1] 1.5
#> 
#> $plant$plant1$codeINN
#> [1] 1
#> 
#> $plant$plant1$codelegume
#> [1] 1
#> 
#> $plant$plant1$stlevdno
#> [1] 150
#> 
#> $plant$plant1$stdnofno
#> [1] 500
#> 
#> $plant$plant1$stfnofvino
#> [1] 300
#> 
#> $plant$plant1$vitno
#> [1] 0.003
#> 
#> $plant$plant1$profnod
#> [1] 40
#> 
#> $plant$plant1$concNnodseuil
#> [1] 1.3
#> 
#> $plant$plant1$concNrac0
#> [1] 1.2
#> 
#> $plant$plant1$concNrac100
#> [1] 0.4
#> 
#> $plant$plant1$tempnod1
#> [1] 0
#> 
#> $plant$plant1$tempnod2
#> [1] 30
#> 
#> $plant$plant1$tempnod3
#> [1] 36
#> 
#> $plant$plant1$tempnod4
#> [1] 50
#> 
#> $plant$plant1$codefixpot
#> [1] 1
#> 
#> $plant$plant1$fixmax
#> [1] 6
#> 
#> $plant$plant1$fixmaxveg
#> [1] 30
#> 
#> $plant$plant1$fixmaxgr
#> [1] 9.5
#> 
#> $plant$plant1$codazofruit
#> [1] 2
#> 
#> $plant$plant1$stadebbchplt
#> [1] 0
#> 
#> $plant$plant1$stadebbchger
#> [1] 5
#> 
#> $plant$plant1$stadebbchlev
#> [1] 9
#> 
#> $plant$plant1$stadebbchamf
#> [1] 35
#> 
#> $plant$plant1$stadebbchlax
#> [1] 55
#> 
#> $plant$plant1$stadebbchsen
#> [1] 85
#> 
#> $plant$plant1$stadebbchflo
#> [1] 65
#> 
#> $plant$plant1$stadebbchdrp
#> [1] 71
#> 
#> $plant$plant1$stadebbchnou
#> [1] -99
#> 
#> $plant$plant1$stadebbchdebdes
#> [1] 75
#> 
#> $plant$plant1$stadebbchmat
#> [1] 89
#> 
#> $plant$plant1$stadebbchrec
#> [1] 99
#> 
#> $plant$plant1$stadebbchfindorm
#> [1] -99
#> 
#> $plant$plant1$nbVariete
#> [1] 16
#> 
#> $plant$plant1$codevar
#>  [1] "DK250"      "Pactol"     "Cherif"     "Furio"      "Dunia"     
#>  [6] "Volga"      "Cecilia"    "DK604"      "Nobilis-DE" "DK300"     
#> [11] "Anjou285"   "DK312"      "DK240"      "banguy"     "magrite"   
#> [16] "clarica"   
#> 
#> $plant$plant1$stlevamf
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        225        253        265        275        285        288        300 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>        300        233        255        242        260        280        280 
#>    magrite    clarica 
#>        280        280 
#> 
#> $plant$plant1$stamflax
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        450        507        530        550        570        577        600 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>        600        467        510        483        520        470        465 
#>    magrite    clarica 
#>        500        545 
#> 
#> $plant$plant1$stlevdrp
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        995       1080       1115       1145       1175       1185       1220 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>       1220       1020       1085       1045       1100       1030       1015 
#>    magrite    clarica 
#>       1080       1125 
#> 
#> $plant$plant1$stflodrp
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        250        250        250        250        250        250        250 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>        250        250        250        250        250        250        250 
#>    magrite    clarica 
#>        250        250 
#> 
#> $plant$plant1$stdrpdes
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        650        650        650        650        650        650        650 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>        650        650        650        650        650        650        650 
#>    magrite    clarica 
#>        650        650 
#> 
#> $plant$plant1$jvc
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>          0          0          0          0          0          0          0 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>          0          0          0          0          0          0          0 
#>    magrite    clarica 
#>          0          0 
#> 
#> $plant$plant1$stdordebour
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>          0          0          0          0          0          0          0 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>          0          0          0          0          0          0          0 
#>    magrite    clarica 
#>          0          0 
#> 
#> $plant$plant1$sensiphot
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>       -999       -999       -999       -999       -999       -999       -999 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>       -999       -999       -999       -999       -999       -999       -999 
#>    magrite    clarica 
#>       -999       -999 
#> 
#> $plant$plant1$phobase
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>       -999       -999       -999       -999       -999       -999       -999 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>       -999       -999       -999       -999       -999       -999       -999 
#>    magrite    clarica 
#>       -999       -999 
#> 
#> $plant$plant1$phosat
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>       -999       -999       -999       -999       -999       -999       -999 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>       -999       -999       -999       -999       -999       -999       -999 
#>    magrite    clarica 
#>       -999       -999 
#> 
#> $plant$plant1$adens
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>      -0.12      -0.12      -0.12      -0.12      -0.12      -0.12      -0.12 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>      -0.12      -0.12      -0.12      -0.12      -0.12      -0.12      -0.12 
#>    magrite    clarica 
#>      -0.12      -0.12 
#> 
#> $plant$plant1$bdens
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>          5          5          5          5          5          5          5 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>          5          5          5          5          5          5          5 
#>    magrite    clarica 
#>          5          5 
#> 
#> $plant$plant1$hautbase
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>          0          0          0          0          0          0          0 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>          0          0          0          0          0          0          0 
#>    magrite    clarica 
#>          0          0 
#> 
#> $plant$plant1$hautmax
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        2.5        2.5        2.5        2.5        2.5        2.5        2.5 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>        2.5        2.5        2.5        2.5        2.5        2.5        2.5 
#>    magrite    clarica 
#>        2.5        2.5 
#> 
#> $plant$plant1$khaut
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        0.7        0.7        0.7        0.7        0.7        0.7        0.7 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>        0.7        0.7        0.7        0.7        0.7        0.7        0.7 
#>    magrite    clarica 
#>        0.7        0.7 
#> 
#> $plant$plant1$durvieF
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        200        200        200        200        200        200        200 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>        200        200        200        200        200        200        200 
#>    magrite    clarica 
#>        200        200 
#> 
#> $plant$plant1$stlaxsen
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        688        730        748        763        778        783        800 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>        800        700        738        713        740        730        750 
#>    magrite    clarica 
#>        730        780 
#> 
#> $plant$plant1$stsenlan
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        272        190        177        197        197        252        220 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>        245        270        222        192        200        300        300 
#>    magrite    clarica 
#>        300        300 
#> 
#> $plant$plant1$dlaimax
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>     0.0016     0.0016     0.0016     0.0016     0.0016     0.0016     0.0016 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>     0.0016     0.0016     0.0016     0.0016     0.0016     0.0016     0.0016 
#>    magrite    clarica 
#>     0.0016     0.0016 
#> 
#> $plant$plant1$dlaimaxbrut
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>     0.0016     0.0016     0.0016     0.0016     0.0016     0.0016     0.0016 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>     0.0016     0.0016     0.0016     0.0016     0.0016     0.0016     0.0016 
#>    magrite    clarica 
#>     0.0016     0.0016 
#> 
#> $plant$plant1$innsen
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>          1          1          1          1          1          1          1 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>          1          1          1          1          1          1          1 
#>    magrite    clarica 
#>          1          1 
#> 
#> $plant$plant1$rapsenturg
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>          0          0          0          0          0          0          0 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>          0          0          0          0          0          0          0 
#>    magrite    clarica 
#>          0          0 
#> 
#> $plant$plant1$extin
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        0.7        0.7        0.7        0.7        0.7        0.7        0.7 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>        0.7        0.7        0.7        0.7        0.7        0.7        0.7 
#>    magrite    clarica 
#>        0.7        0.7 
#> 
#> $plant$plant1$ktrou
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>          1          1          1          1          1          1          1 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>          1          1          1          1          1          1          1 
#>    magrite    clarica 
#>          1          1 
#> 
#> $plant$plant1$temin
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>          8          8          8          8          8          8          8 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>          8          8          8          8          8          8          8 
#>    magrite    clarica 
#>          8          8 
#> 
#> $plant$plant1$teopt
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>         24         24         24         24         24         24         24 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>         24         24         24         24         24         24         24 
#>    magrite    clarica 
#>         24         24 
#> 
#> $plant$plant1$slamax
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        250        250        250        250        250        250        250 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>        250        250        250        250        250        250        250 
#>    magrite    clarica 
#>        250        250 
#> 
#> $plant$plant1$tigefeuil
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>       0.25       0.25       0.25       0.25       0.25       0.25       0.25 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>       0.25       0.25       0.25       0.25       0.25       0.25       0.25 
#>    magrite    clarica 
#>       0.25       0.25 
#> 
#> $plant$plant1$pgrainmaxi
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>      0.313      0.296      0.330      0.330      0.330      0.322      0.348 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>      0.330      0.313      0.322      0.260      0.313      0.260      0.280 
#>    magrite    clarica 
#>      0.310      0.320 
#> 
#> $plant$plant1$vitpropsucre
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>          0          0          0          0          0          0          0 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>          0          0          0          0          0          0          0 
#>    magrite    clarica 
#>          0          0 
#> 
#> $plant$plant1$vitprophuile
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>          0          0          0          0          0          0          0 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>          0          0          0          0          0          0          0 
#>    magrite    clarica 
#>          0          0 
#> 
#> $plant$plant1$vitirazo
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>      0.011      0.011      0.011      0.011      0.011      0.011      0.011 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>      0.011      0.011      0.011      0.011      0.011      0.011      0.011 
#>    magrite    clarica 
#>      0.011      0.011 
#> 
#> $plant$plant1$deshydbase
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>      0.008      0.008      0.008      0.008      0.008      0.008      0.008 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>      0.008      0.008      0.008      0.008      0.008      0.008      0.008 
#>    magrite    clarica 
#>      0.008      0.008 
#> 
#> $plant$plant1$nbjgrain
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>         20         20         20         20         20         20         20 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>         20         20         20         20         20         20         20 
#>    magrite    clarica 
#>         20         20 
#> 
#> $plant$plant1$nbgrmin
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>       1500       1500       1500       1500       1500       1500       1500 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>       1500       1500       1500       1500       1500       1500       1500 
#>    magrite    clarica 
#>       1500       1500 
#> 
#> $plant$plant1$nbgrmax
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>       4000       4200       4500       4500       4500       5000       5000 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>       5000       4500       4200       4500       4200       4500       4500 
#>    magrite    clarica 
#>       4500       4500 
#> 
#> $plant$plant1$stdrpmat
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        640        600        605        640        655        715        700 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>        725        650        640        585        620        750        780 
#>    magrite    clarica 
#>        730        790 
#> 
#> $plant$plant1$vitircarb
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>     0.0103     0.0103     0.0103     0.0103     0.0103     0.0103     0.0103 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>     0.0103     0.0103     0.0103     0.0103     0.0103     0.0103     0.0103 
#>    magrite    clarica 
#>     0.0103     0.0103 
#> 
#> $plant$plant1$vitircarbT
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>     0.0011     0.0011     0.0011     0.0011     0.0011     0.0011     0.0011 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>     0.0011     0.0011     0.0011     0.0011     0.0011     0.0011     0.0011 
#>    magrite    clarica 
#>     0.0011     0.0011 
#> 
#> $plant$plant1$afruitpot
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>          0          0          0          0          0          0          0 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>          0          0          0          0          0          0          0 
#>    magrite    clarica 
#>          0          0 
#> 
#> $plant$plant1$dureefruit
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>          0          0          0          0          0          0          0 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>          0          0          0          0          0          0          0 
#>    magrite    clarica 
#>          0          0 
#> 
#> $plant$plant1$stdrpnou
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>       -999       -999       -999       -999       -999       -999       -999 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>       -999       -999       -999       -999       -999       -999       -999 
#>    magrite    clarica 
#>       -999       -999 
#> 
#> $plant$plant1$nbinflo
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>       -999       -999       -999       -999       -999       -999       -999 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>       -999       -999       -999       -999       -999       -999       -999 
#>    magrite    clarica 
#>       -999       -999 
#> 
#> $plant$plant1$inflomax
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>       -999       -999       -999       -999       -999       -999       -999 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>       -999       -999       -999       -999       -999       -999       -999 
#>    magrite    clarica 
#>       -999       -999 
#> 
#> $plant$plant1$pentinflores
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>       -999       -999       -999       -999       -999       -999       -999 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>       -999       -999       -999       -999       -999       -999       -999 
#>    magrite    clarica 
#>       -999       -999 
#> 
#> $plant$plant1$croirac
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>       0.15       0.15       0.15       0.15       0.15       0.15       0.15 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>       0.15       0.15       0.15       0.15       0.15       0.15       0.15 
#>    magrite    clarica 
#>       0.15       0.15 
#> 
#> $plant$plant1$tgellev10
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>         -1         -1         -1         -1         -1         -1         -1 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>         -1         -1         -1         -1         -1         -1         -1 
#>    magrite    clarica 
#>         -1         -1 
#> 
#> $plant$plant1$tgeljuv10
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>         -1         -1         -1         -1         -1         -1         -1 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>         -1         -1         -1         -1         -1         -1         -1 
#>    magrite    clarica 
#>         -1         -1 
#> 
#> $plant$plant1$tgelveg10
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>         -1         -1         -1         -1         -1         -1         -1 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>         -1         -1         -1         -1         -1         -1         -1 
#>    magrite    clarica 
#>         -1         -1 
#> 
#> $plant$plant1$psisto
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>         12         12         12         12         12         12         12 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>         12         12         12         12         12         12         12 
#>    magrite    clarica 
#>         12         12 
#> 
#> $plant$plant1$psiturg
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>          5          5          5          5          5          5          5 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>          5          5          5          5          5          5          5 
#>    magrite    clarica 
#>          5          5 
#> 
#> $plant$plant1$swfacmin
#>      DK250     Pactol     Cherif      Furio      Dunia      Volga    Cecilia 
#>        0.1        0.1        0.1        0.1        0.1        0.1        0.1 
#>      DK604 Nobilis-DE      DK300   Anjou285      DK312      DK240     banguy 
#>        0.1        0.1        0.1        0.1        0.1        0.1        0.1 
#>    magrite    clarica 
#>        0.1        0.1 
#> 
#> 
#> 
#> $soil
#> $soil$nbcouchessol_max
#> [1] 1000
#> 
#> $soil$numsol
#> [1] 1
#> 
#> $soil$typsol
#> [1] "solmais"
#> 
#> $soil$argi
#> [1] 22
#> 
#> $soil$Norg
#> [1] 0.12
#> 
#> $soil$profhum
#> [1] 35
#> 
#> $soil$calc
#> [1] 1
#> 
#> $soil$pH
#> [1] 7
#> 
#> $soil$concseuil
#> [1] 0
#> 
#> $soil$albedo
#> [1] 0.2
#> 
#> $soil$q0
#> [1] 0
#> 
#> $soil$ruisolnu
#> [1] 0
#> 
#> $soil$obstarac
#> [1] 200
#> 
#> $soil$pluiebat
#> [1] 50
#> 
#> $soil$mulchbat
#> [1] 0.5
#> 
#> $soil$zesx
#> [1] 60
#> 
#> $soil$cfes
#> [1] 5
#> 
#> $soil$z0solnu
#> [1] 0.01
#> 
#> $soil$CsurNsol
#> [1] 0
#> 
#> $soil$finert
#> [1] 0.65
#> 
#> $soil$penterui
#> [1] 0.33
#> 
#> $soil$codecailloux
#> [1] 2
#> 
#> $soil$codemacropor
#> [1] 2
#> 
#> $soil$codefente
#> [1] 2
#> 
#> $soil$codrainage
#> [1] 2
#> 
#> $soil$coderemontcap
#> [1] 2
#> 
#> $soil$codenitrif
#> [1] 2
#> 
#> $soil$codedenit
#> [1] 2
#> 
#> $soil$profimper
#> [1] 10
#> 
#> $soil$ecartdrain
#> [1] 0
#> 
#> $soil$ksol
#> [1] 0
#> 
#> $soil$profdrain
#> [1] 0
#> 
#> $soil$capiljour
#> [1] 1
#> 
#> $soil$humcapil
#> [1] 10
#> 
#> $soil$profdenit
#> [1] 20
#> 
#> $soil$vpotdenit
#> [1] 2
#> 
#> $soil$epc
#> [1] 25 30 35 30  0
#> 
#> $soil$hccf
#> [1] 22.5 25.4 25.8 28.7  0.0
#> 
#> $soil$hminf
#> [1]  8.5 11.6 15.3 19.0  0.0
#> 
#> $soil$DAF
#> [1] 1.16 1.29 1.15 0.91 0.00
#> 
#> $soil$cailloux
#> [1] 0 0 0 0 0
#> 
#> $soil$typecailloux
#> [1] 1 1 1 1 1
#> 
#> $soil$infil
#> [1] 50 50 50 50 50
#> 
#> $soil$epd
#> [1] 10 10 10 10 10
#> 
#> 
#> $station
#> $station$zr
#> [1] 2.5
#> 
#> $station$NH3ref
#> [1] 0
#> 
#> $station$concrr
#> [1] 0.02
#> 
#> $station$latitude
#> [1] 45
#> 
#> $station$patm
#> [1] 1000
#> 
#> $station$aclim
#> [1] 20
#> 
#> $station$codeetp
#> [1] 1
#> 
#> $station$alphapt
#> [1] 1.26
#> 
#> $station$codeclichange
#> [1] 1
#> 
#> $station$codaltitude
#> [1] 1
#> 
#> $station$altistation
#> [1] 440
#> 
#> $station$altisimul
#> [1] 800
#> 
#> $station$gradtn
#> [1] -0.5
#> 
#> $station$gradtx
#> [1] -0.55
#> 
#> $station$altinversion
#> [1] 500
#> 
#> $station$gradtninv
#> [1] 1.3
#> 
#> $station$cielclair
#> [1] 0.8
#> 
#> $station$codadret
#> [1] 1
#> 
#> $station$ombragetx
#> [1] -1.4
#> 
#> $station$ra
#> [1] 50
#> 
#> $station$albveg
#> [1] 0.23
#> 
#> $station$aangst
#> [1] 0.18
#> 
#> $station$bangst
#> [1] 0.62
#> 
#> $station$corecTrosee
#> [1] 1
#> 
#> $station$codecaltemp
#> [1] 1
#> 
#> $station$codernet
#> [1] 2
#> 
#> $station$coefdevil
#> [1] 0.7
#> 
#> $station$aks
#> [1] 6
#> 
#> $station$bks
#> [1] 0.5
#> 
#> $station$cvent
#> [1] 0.16
#> 
#> $station$phiv0
#> [1] 0.004
#> 
#> $station$coefrnet
#> [1] 0.59
#> 
#> $station$codemodlsnow
#> [1] 3
#> 
#> $station$tsmax
#> [1] -2
#> 
#> $station$trmax
#> [1] 1
#> 
#> $station$DKmax
#> [1] 1.5
#> 
#> $station$Kmin
#> [1] 2
#> 
#> $station$Tmf
#> [1] 0.5
#> 
#> $station$SWrf
#> [1] 0.01
#> 
#> $station$Pns
#> [1] 100
#> 
#> $station$E
#> [1] 0.02
#> 
#> $station$prof
#> [1] 10
#> 
#> $station$tminseuil
#> [1] -0.5
#> 
#> $station$tmaxseuil
#> [1] 0
#> 
#> 
#> $tmp
#> $tmp$codepluiepoquet
#> [1] 2
#> 
#> $tmp$nbjoursrrversirrig
#> [1] 15
#> 
#> $tmp$codecalferti
#> [1] 2
#> 
#> $tmp$ratiolN
#> [1] 0
#> 
#> $tmp$dosimxN
#> [1] 40
#> 
#> $tmp$codetesthumN
#> [1] 1
#> 
#> $tmp$codeNmindec
#> [1] 2
#> 
#> $tmp$rapNmindec
#> [1] 0.001
#> 
#> $tmp$fNmindecmin
#> [1] 0.1
#> 
#> $tmp$codetrosee
#> [1] 1
#> 
#> $tmp$codeSWDRH
#> [1] 2
#> 
#> $tmp$option_pature
#> [1] 2
#> 
#> $tmp$coderes_pature
#> [1] 3
#> 
#> $tmp$pertes_restit_ext
#> [1] 0.2
#> 
#> $tmp$Crespc_pature
#> [1] 7.4
#> 
#> $tmp$Nminres_pature
#> [1] 0.45
#> 
#> $tmp$eaures_pature
#> [1] 87
#> 
#> $tmp$coef_calcul_qres
#> [1] 7.53
#> 
#> $tmp$engrais_pature
#> [1] 3
#> 
#> $tmp$coef_calcul_doseN
#> [1] 16.25
#> 
#> $tmp$code_CsurNsol_dynamic
#> [1] 2
#> 
#> $tmp$humirac
#> [1] 1
#> 
#> $tmp$code_ISOP
#> [1] 2
#> 
#> $tmp$code_pct_legume
#> [1] 2
#> 
#> $tmp$pct_legum
#> [1] 0.5
#> 
#> 
get_param_txt(workspace = path, param = "durvieF", plant_id = 1)
#> $plant
#> $plant$plant1
#> $plant$plant1$durvieF
#> Furio 
#>   200 
#> 
#> 
#> 
get_param_txt(
  workspace = path, param = "durvieF", plant_id = 1,
  variety = varieties
)
#> $plant
#> $plant$plant1
#> $plant$plant1$durvieF
#>  Pactol  Cherif   Furio   Dunia   Volga Cecilia 
#>     200     200     200     200     200     200 
#> 
#> 
#> 

# Get parameters for specific interventions or soil layers
get_param_txt(workspace = path, param = "amount", value_id = c(1, 3))
#> $tec
#> $tec$plant1
#> $tec$plant1$amount
#> [1] 20 29
#> 
#> 
#> 
get_param_txt(workspace = path, param = "Hinitf", value_id = c(1, 3))
#> $ini
#> $ini$Hinitf
#> [1] 23.5 23.9
#> 
#> 
get_param_txt(workspace = path, param = "epc", value_id = c(1, 3))
#> $soil
#> $soil$epc
#> [1] 25 35
#> 
#> 

if (FALSE) { # \dontrun{
# Read the initialisation file (ficini.txt):
library(SticsRFiles)
path <- file.path(get_examples_path(file_type = "txt"), "ficini.txt")
get_ini_txt(path)

# Read the tec file directly:

# First, get the parameters from the tmp file:
tmp <- get_tmp_txt(file = file.path(
  get_examples_path(file_type = "txt"),
  "tempoparv6.sti"
))
several_fert <- ifelse(tmp$option_engrais_multiple == 1, TRUE, FALSE)
several_thin <- ifelse(tmp$option_thinning == 1, TRUE, FALSE)
is_pasture <- ifelse(tmp$option_pature == 1, TRUE, FALSE)

# Then, get the technical parameters:
get_tec_txt(
  file = file.path(get_examples_path(file_type = "txt"), "fictec1.txt"),
  several_fert = several_fert, several_thin = several_thin,
  is_pasture = is_pasture
)
} # }