Extracting data from the Stics report file

Extracting data from the Stics report file

get_report_results(
  workspace,
  file_name = "mod_rapport.sti",
  usm = NULL,
  var_list = NULL,
  usm_name = lifecycle::deprecated()
)

Arguments

workspace

Path of the directory containing the Stics report file to read.

file_name

A report file name among "mod_rapport.sti" (default), "mod_rapportA.sti", "mod_rapportP.sti"

usm

Vector of USM names. Optional, if not provided, the function returns the results for all USMs.

var_list

vector of output variables names to filter (optional, see get_var_info() to get the names of the variables)

usm_name

usm_name is no longer supported, use usm instead.

Value

A data.frame

Details

The data may be filtered using usm_name vector of usm names and and/or var_list vector of variables names. In the returned data.frame, variables names respect the same syntax as in the get_sim output.

Examples

if (FALSE) {
path <- get_examples_path(file_type = "sti")
get_report_results(workspace = path)

get_report_results(workspace = path, usm = c("DurumWheat", "grass"))

get_report_results(workspace = path, var_list = c("masec(n)", "QNplante"))

get_report_results(workspace = path, usm = c("DurumWheat", "grass"))

get_report_results(workspace = path)

get_report_results(workspace = path, file_name = "mod_rapportA.sti")
}