Find STICS output variable names and description

Helper function that returns names and descriptions of STICS output variables from a partial name and/or descriptive keywords.

get_var_info(
  var = NULL,
  keyword = NULL,
  stics_version = "latest",
  version = lifecycle::deprecated()
)

Arguments

var

Vector of variable names (or partial names). Optional, if not provided, the function returns information for all variables.

keyword

Search by keyword instead of variable name (search in the name and description field)

stics_version

Name of the Stics version. Optional, can be used to search parameters information relative to a specific Stics version. By default the latest version returned by get_stics_versions_compat() is used.

version

version is no longer supported, use stics_version instead.

Details

The function understand regex as input.

See also

Examples

if (FALSE) {
# Find by variable name (fuzzy search):
SticsRFiles::get_var_info("lai")

# Find by keyword (fuzzy search in variable name and description):
SticsRFiles::get_var_info(keyword = "lai")

# Find for a particular version:
SticsRFiles::get_var_info("lai", stics_version = "V9.0")
}