Extracting a data.frame of parameters values for an xml file or a set of (whatever the xml file kind)

Extracting a data.frame of parameters values for an xml file or a set of (whatever the xml file kind)

get_xml_files_param_df(
  file_path,
  select = NULL,
  name = NULL,
  param_names = NULL,
  wide_shape = FALSE
)

Arguments

file_path
A file path or a vector of

select
node name or attribute name to use for selection (optional, default to no selection)

name
value used for select (optional)

param_names
vector of parameters names (optional)

wide_shape
Optional logical for keeping the long data.frame format FALSE, default) or converting it to a wider format one (TRUE)

Value

A data.frame with name, type, param, id and value columns

Details

An extract of an example of data.frame for the 2 formats

Default one

name	type	param	id	value

SugarCane	usms	datedebut	NA	286

SugarCane	usms	datefin	NA	650

SugarCane	usms	finit	NA	canne_ini.xml

SugarCane	usms	nomsol	NA	solcanne

SugarCane	usms	fstation	NA	climcanj_sta.xml

...	...	...	...	...

...	...	...	...	...

param_gen.xml	fichierpar	Qmulchdec	17	0

param_gen.xml	fichierpar	Qmulchdec	18	0

param_gen.xml	fichierpar	Qmulchdec	19	0

param_gen.xml	fichierpar	Qmulchdec	20	0

param_gen.xml	fichierpar	Qmulchdec	21	0

name: a file name or an usm or soil name type: type of file param: param name id: NA for scalar, integer id for vectors parameters value: parameter value

wide format

name	type	datedebut	datefin	finit	nomsol	fstation	fclim1	fclim2

SugarCane	usms	286	650	canne_ini.xml	solcanne	climcanj_sta.xml	climcanj.1998	climcanj.1999

potato	usms	91	250	patate_ini.xml	solpatate	climpdtj_sta.xml	climpdtj.1997	climpdtj.1997

banana	usms	30	300	banane_ini.xml	solbanane	climbanj_sta.xml	climbanj.1996	climbanj.1996

sorghum	usms	112	360	sorgho_ini.xml	solsorgho	climsorj_sta.xml	climsorj.1996	climsorj.1996

barley	usms	88	196	orge_ini.xml	solorge	climorgj_sta.xml	climorgj.2002	climorgj.2002

name: a file name or an usm or soil name type: type of file all other columns names correspond to parameter names, with indices as suffix for multiple values

Examples

if (FALSE) {
dir_path <- get_examples_path("xml")
get_xml_files_param_df(file_path = file.path(dir_path,"sols.xml"))

get_xml_files_param_df(file_path = file.path(dir_path,"sols.xml"),
select = "sol", c("solcanne", "solble"))

files_list <- file.path(dir_path, c("sols.xml", "usms.xml", "param_gen.xml" ))
get_xml_files_param_df(file_path = files_list)

}