Set (replace) STICS input file parameters

Replace or set an input parameter from a pre-existing STICS input file.

set_param_txt(dirpath = getwd(), param, value, add = F, plant = 1)

set_usm_txt(filepath = "new_travail.usm", param, value, add = F)

set_station_txt(filepath = "station.txt", param, value, add = F)

set_ini_txt(filepath = "ficini.txt", param, value, add = F)

set_general_txt(filepath = "tempopar.sti", param, value, add = F)

set_tmp_txt(filepath = "tempoparv6.sti", param, value, add = F)

set_plant_txt(filepath = "ficplt1.txt", param, value, add = F, variety = NULL)

set_tec_txt(filepath = "fictec1.txt", param, value, add = F)

set_soil_txt(filepath = "param.sol", param, value)

set_out_var_txt(filepath = "var.mod", vars = c("lai(n)", "masec(n)"), add = F)

Arguments

dirpath
USM directory path

param
Parameter name

value
New parameter value

add
Boolean. Append input to existing file (add to the list)

plant
Plant index. Optional, only for plant or technical parameters

filepath
Path to the parameter file

variety
Integer. The plant variety to get the parameter from.

vars
Vector of variable names for STICS output requirements

Details

The plant parameter can be either equal to 1, 2 for the associated plant in the case of intercrop, or c(1,2) for both Princiapl and associated plants. all_out_var is a helper function that returns all possible output variables.

Note

set_out_var_txt is not used by set_param_txt. To replace the output variables required from STICS, please directly call set_out_var_txt.

Examples

if (FALSE) {
# Replace the interrow distance parameter to 0.01:

library(SticsRFiles)
set_param_txt(dirpath = "stics_usm/usm_1", param= "interrang", value= 0.01)

}