Setting parameter values into xml files

Setting parameter values for a parameter or a vector of and with a parameters values vector

set_param_xml(
  xml_file,
  param_names,
  param_values,
  out_path = NULL,
  out_file = NULL,
  select = NULL,
  value = NULL,
  overwrite = FALSE,
  ...
)

Arguments

xml_file
an xml file path

param_names
parameter names vector (i.e.: option parameter name, parameter name, other kinds in ini and some tec parameters not taken into account for the moment)

param_values
parameter(s) values or a list of

out_path
a directory path

out_file
an xml file name

select
node name or attribute name name to use for selection (optional)

value
value used for select (optional)

overwrite
Logical TRUE for overwriting the output file, FALSE otherwise (default)

...
To pass some other arguments

Value

A logical value TRUE for operation success, FALSE otherwise

Examples

if (FALSE) {
xml_path = system.file("extdata/xml/examples/V9.1/sols.xml", package = "SticsRFiles")
file.copy(xml_path, getwd())
get_param_xml("sols.xml", "argi")

# Setting all "argi" values to 50
set_param_xml("sols.xml", "argi", 50)

get_param_xml("sols.xml", "argi")

# Setting a specific value to "argi" for "solcanne" soil
set_param_xml("sols.xml", "argi",56,
select = "sol", value = "solcanne")

get_param_xml("sols.xml", "argi",
select = "sol", value = "solcanne")

# Setting a specific values to 2 parameters "argi" and "norg" for "solcanne" soil
set_param_xml("sols.xml", c("argi", "norg"),list(100,150),
select = "sol", value = "solcanne")

get_param_xml("sols.xml", c("argi", "norg"),
select = "sol", value = "solcanne")

}